Feat/cosmosdb client encryption (#30739)
@azure/cosmos

This PR adds diagnostics, and test cases for Client Side Encryption

there are more than one possible design, why was the one in this PR
chosen?

request PRs)_

- [ ] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)

---------
