+++ sdk/cosmosdb/cosmos/src/CosmosDiagnostics.ts
@@ -165,6 +165,15 @@ export interface FailedRequestAttemptDiagnostic {
 }
 
 /**
+ * Represents the diagnostics information for encryption operations.
+ */
+export interface EncryptionDiagnostics {
+  encryptContent: { [key: string]: any };
+  decryptContent: { [key: string]: any };
+  processingDurationInMs: number;
+}
+
+/**
  * This is enum for Type of Metadata lookups possible.
  */
 export enum MetadataLookUpType {
@@ -211,6 +220,10 @@ export type ClientSideRequestStatistics = {
    * This is the cumulated Response Payload Length n bytes, this includes metadata calls along with the main operation.
    */
   totalResponsePayloadLengthInBytes: number;
+  /**
+   * This field captures diagnostic information for encryption/decryption happened during CRUD operation if encryption is enabled.
+   */
+  encryptionDiagnostics?: EncryptionDiagnostics;
 };
 
 export function getRootNode(node: DiagnosticNodeInternal): DiagnosticNodeInternal {
