+++ sdk/cosmosdb/cosmos/src/client/ChangeFeed/changeFeedUtils.ts
@@ -13,6 +13,9 @@ import { QueryRange } from "../../routing";
 import { FeedRangeInternal } from "./FeedRange";
 import { hashV2PartitionKey } from "../../utils/hashing/v2";
 import { PartitionKeyInternal } from "../../documents/PartitionKeyInternal";
+import { DiagnosticNodeInternal } from "../../diagnostics/DiagnosticNodeInternal";
+import { EncryptionProcessor } from "../../encryption";
+import { ChangeFeedMode } from "./ChangeFeedMode";
 
 /**
  * @hidden
@@ -85,9 +88,9 @@ export function isEpkRange(obj: unknown): boolean {
     typeof obj.minInclusive === "string" &&
     typeof obj.maxExclusive === "string" &&
     obj.minInclusive >=
-      Constants.EffectivePartitionKeyConstants.MinimumInclusiveEffectivePartitionKey &&
+    Constants.EffectivePartitionKeyConstants.MinimumInclusiveEffectivePartitionKey &&
     obj.maxExclusive <=
-      Constants.EffectivePartitionKeyConstants.MaximumExclusiveEffectivePartitionKey &&
+    Constants.EffectivePartitionKeyConstants.MaximumExclusiveEffectivePartitionKey &&
     obj.maxExclusive > obj.minInclusive
   );
 }
@@ -152,4 +155,27 @@ export function getEffectivePartitionKeyForMultiHashPartitioning(
 ): string {
   const hashArray = partitionKeyInternal.map((item) => hashV2PartitionKey([item]));
   return hashArray.join("");
+}
+
+/**
+ * @hidden
+ */
+export async function decryptChangeFeedResponse(
+  result: any,
+  diagnosticNode: DiagnosticNodeInternal,
+  changeFeedMode: ChangeFeedMode,
+  encryptionProcessor: EncryptionProcessor,
+): Promise<void> {
+  for (let item of result.result) {
+    if (changeFeedMode === ChangeFeedMode.AllVersionsAndDeletes) {
+      if ("current" in item && item.current !== null) {
+        item.current = await encryptionProcessor.decrypt(item.current, diagnosticNode);
+      }
+      if ("previous" in item && item.previous !== null) {
+        item.previous = await encryptionProcessor.decrypt(item.previous, diagnosticNode);
+      }
+    } else {
+      item = await encryptionProcessor.decrypt(item, diagnosticNode);
+    }
+  }
 }
