+++ sdk/cosmosdb/cosmos/src/client/Container/Container.ts
@@ -117,6 +117,10 @@ export class Container {
    */
   public _rid: string;
   /**
+   * @internal
+   */
+  public isEncryptionInitialized: boolean = false;
+  /**
    * Returns a container instance. Note: You should get this from `database.container(id)`, rather than creating your own object.
    * @param database - The parent {@link Database}.
    * @param id - The id of the given container.
@@ -129,15 +133,16 @@ export class Container {
     private encryptionManager?: EncryptionManager,
     _rid?: string,
   ) {
+    this._rid = _rid;
     if (this.clientContext.enableEncryption) {
       this.encryptionProcessor = new EncryptionProcessor(
         this.id,
-        this.database.id,
+        this._rid,
+        this.database,
         this.clientContext,
         this.encryptionManager,
       );
     }
-    this._rid = _rid;
   }
 
   /**
@@ -379,42 +384,40 @@ export class Container {
       const id = getIdFromLink(this.url);
       path = path + "/operations/partitionkeydelete";
       if (this.clientContext.enableEncryption) {
+        if (!this.isEncryptionInitialized) {
+          await this.initializeEncryption();
+        }
+        this.encryptionProcessor.containerRid = this._rid;
+        options = options || {};
+        options.containerRid = this._rid;
         const partitionKeyInternal = convertToInternalPartitionKey(partitionKey);
         partitionKey =
           await this.encryptionProcessor.getEncryptedPartitionKeyValue(partitionKeyInternal);
       }
-      const response = await this.clientContext.delete<ContainerDefinition>({
-        path,
-        resourceType: ResourceType.container,
-        resourceId: id,
-        options,
-        partitionKey: partitionKey,
-        method: HTTPMethod.post,
-        diagnosticNode,
-      });
+      let response: Response<any>;
+      try {
+        response = await this.clientContext.delete<ContainerDefinition>({
+          path,
+          resourceType: ResourceType.container,
+          resourceId: id,
+          options,
+          partitionKey: partitionKey,
+          method: HTTPMethod.post,
+          diagnosticNode,
+        });
+      } catch (error) {
+        if (this.clientContext.enableEncryption) {
+          await this.throwIfRequestNeedsARetryPostPolicyRefresh(error);
+        }
+        throw error;
+      }
+
       return new ContainerResponse(
         response.result,
         response.headers,
         response.code,
         this,
         getEmptyCosmosDiagnostics(),
-      );
-    }, this.clientContext);
-  }
-
-  private async refreshRidAndPolicy(): Promise<void> {
-    await withDiagnostics(async (diagnosticNode: DiagnosticNodeInternal) => {
-      const readResponse = await this.readInternal(diagnosticNode);
-      this._rid = readResponse.resource._rid;
-      const clientEncryptionPolicy = readResponse.resource.clientEncryptionPolicy;
-      if (!clientEncryptionPolicy) return;
-      const partitionKeyPaths = readResponse.resource.partitionKey.paths;
-      const key = this.database.id + "/" + this.id;
-      await this.encryptionManager.encryptionSettingsCache.createAndSetEncryptionSettings(
-        key,
-        this._rid,
-        partitionKeyPaths,
-        clientEncryptionPolicy,
       );
     }, this.clientContext);
   }
@@ -432,7 +435,10 @@ export class Container {
         const clientEncryptionPolicy = readResponse.resource.clientEncryptionPolicy;
         if (!clientEncryptionPolicy) return;
         const partitionKeyPaths = readResponse.resource.partitionKey.paths;
-        const key = this.database.id + "/" + this.id;
+        const { resource: databaseDefinition } = await this.database.read();
+        this.database._rid = databaseDefinition._rid;
+
+        const key = this.database._rid + "/" + this._rid;
 
         await this.encryptionManager.encryptionSettingsCache.createAndSetEncryptionSettings(
           key,
@@ -449,19 +455,25 @@ export class Container {
         for (const clientEncryptionKeyId of clientEncryptionKeyIds) {
           const res = await this.database.readClientEncryptionKey(clientEncryptionKeyId);
           const encryptionKeyProperties = res.clientEncryptionKeyProperties;
-          const key = this.database.id + "/" + clientEncryptionKeyId;
+          const key = this.database._rid + "/" + clientEncryptionKeyId;
 
           this.encryptionManager.clientEncryptionKeyPropertiesCache.setClientEncryptionKeyProperties(
             key,
             encryptionKeyProperties,
           );
         }
+        this.isEncryptionInitialized = true;
       }, this.clientContext);
     }
   }
-
-  async ThrowIfRequestNeedsARetryPostPolicyRefresh(errorResponse: any): Promise<void> {
-    const key = this.database.id + "/" + this.id;
+  /**
+   * This function handles the scenario where a container is deleted(say from different Client) and recreated with same Id but with different client encryption policy.
+   * The idea is to have the container Rid cached and sent out as part of RequestOptions with Container Rid set in "x-ms-cosmos-intended-collection-rid" header.
+   * So, when the container being referenced here gets recreated we would end up with a stale encryption settings and container Rid and this would result in BadRequest (and a substatus 1024).
+   * This would allow us to refresh the encryption settings and Container Rid, on the premise that the container recreated could possibly be configured with a new encryption policy.
+   */
+  async throwIfRequestNeedsARetryPostPolicyRefresh(errorResponse: any): Promise<void> {
+    const key = this.database._rid + "/" + this._rid;
     const encryptionSetting =
       this.encryptionManager.encryptionSettingsCache.getEncryptionSettings(key);
     const subStatusCode = errorResponse.headers[Constants.HttpHeaders.SubStatus];
@@ -496,9 +508,10 @@ export class Container {
       if (currentContainerRid === updatedContainerRid) {
         return;
       }
-      await this.refreshRidAndPolicy();
+      await this.initializeEncryption();
       throw new ErrorResponse(
-        "Operation has failed due to a possible mismatch in Client Encryption Policy configured on the container. Retrying may fix the issue.",
+        "Operation has failed due to a possible mismatch in Client Encryption Policy configured on the container. Retrying may fix the issue. Please refer to https://aka.ms/CosmosClientEncryption for more details." +
+          errorResponse.message,
       );
     }
   }
