+++ sdk/cosmosdb/cosmos/src/client/Item/Items.ts
@@ -5,7 +5,7 @@ import { ChangeFeedIterator } from "../../ChangeFeedIterator";
 import type { ChangeFeedOptions } from "../../ChangeFeedOptions";
 import type { ClientContext } from "../../ClientContext";
 import {
-  addContainerRid,
+  Constants,
   copyObject,
   getIdFromLink,
   getPathFromLink,
@@ -201,11 +201,12 @@ export class Items {
   }
 
   private async buildSqlQuerySpec(encryptionSqlQuerySpec: SqlQuerySpec): Promise<SqlQuerySpec> {
-    const encryptionParameters = encryptionSqlQuerySpec.parameters as EncryptionSqlParameter[];
+    let encryptionParameters = encryptionSqlQuerySpec.parameters as EncryptionSqlParameter[];
     const sqlQuerySpec: SqlQuerySpec = {
       query: encryptionSqlQuerySpec.query,
       parameters: [],
     };
+    encryptionParameters = copyObject(encryptionParameters);
     for (const parameter of encryptionParameters) {
       const value = await this.container.encryptionProcessor.encryptQueryParameter(
         parameter.path,
@@ -398,9 +399,12 @@ export class Items {
       );
       let partitionKey = extractPartitionKeys(body, partitionKeyDefinition);
       if (this.clientContext.enableEncryption) {
-        addContainerRid(this.container);
+        if (!this.container.isEncryptionInitialized) {
+          await this.container.initializeEncryption();
+        }
+        this.container.encryptionProcessor.containerRid = this.container._rid;
         body = copyObject(body);
-        body = await this.container.encryptionProcessor.encrypt(body);
+        body = await this.container.encryptionProcessor.encrypt(body, diagnosticNode);
         options.containerRid = this.container._rid;
         partitionKey = extractPartitionKeys(body, partitionKeyDefinition);
       }
@@ -423,13 +427,17 @@ export class Items {
         });
       } catch (error: any) {
         if (this.clientContext.enableEncryption) {
-          await this.container.ThrowIfRequestNeedsARetryPostPolicyRefresh(error);
+          // Todo: internally retry post policy refresh
+          await this.container.throwIfRequestNeedsARetryPostPolicyRefresh(error);
         }
         throw error;
       }
 
       if (this.clientContext.enableEncryption) {
-        response.result = await this.container.encryptionProcessor.decrypt(response.result);
+        response.result = await this.container.encryptionProcessor.decrypt(
+          response.result,
+          diagnosticNode,
+        );
         partitionKey = extractPartitionKeys(response.result, partitionKeyDefinition);
       }
       const ref = new Item(
@@ -496,9 +504,12 @@ export class Items {
       if (this.clientContext.enableEncryption) {
         body = copyObject(body);
         options = options || {};
-        addContainerRid(this.container);
+        if (!this.container.isEncryptionInitialized) {
+          await this.container.initializeEncryption();
+        }
+        this.container.encryptionProcessor.containerRid = this.container._rid;
         options.containerRid = this.container._rid;
-        body = await this.container.encryptionProcessor.encrypt(body);
+        body = await this.container.encryptionProcessor.encrypt(body, diagnosticNode);
         partitionKey = extractPartitionKeys(body, partitionKeyDefinition);
       }
 
@@ -522,12 +533,15 @@ export class Items {
         });
       } catch (error: any) {
         if (this.clientContext.enableEncryption) {
-          await this.container.ThrowIfRequestNeedsARetryPostPolicyRefresh(error);
+          await this.container.throwIfRequestNeedsARetryPostPolicyRefresh(error);
         }
         throw error;
       }
       if (this.clientContext.enableEncryption) {
-        response.result = await this.container.encryptionProcessor.decrypt(response.result);
+        response.result = await this.container.encryptionProcessor.decrypt(
+          response.result,
+          diagnosticNode,
+        );
         partitionKey = extractPartitionKeys(response.result, partitionKeyDefinition);
       }
 
@@ -594,9 +608,12 @@ export class Items {
       if (this.clientContext.enableEncryption) {
         operations = copyObject(operations);
         options = options || {};
-        addContainerRid(this.container);
+        if (!this.container.isEncryptionInitialized) {
+          await this.container.initializeEncryption();
+        }
+        this.container.encryptionProcessor.containerRid = this.container._rid;
         options.containerRid = this.container._rid;
-        operations = await this.bulkBatchEncryptionHelper(operations);
+        operations = await this.bulkBatchEncryptionHelper(operations, diagnosticNode);
       }
 
       const batches: Batch[] = partitionKeyRanges.map((keyRange: PartitionKeyRange) => {
@@ -665,16 +682,23 @@ export class Items {
           DiagnosticNodeType.BATCH_REQUEST,
         );
         if (this.clientContext.enableEncryption) {
+          diagnosticNode.beginEncryptionDiagnostics(
+            Constants.Encryption.DiagnosticsDecryptOperation,
+          );
           for (const result of response.result) {
             result.resourceBody = await this.container.encryptionProcessor.decrypt(
               result.resourceBody,
             );
           }
+          diagnosticNode.endEncryptionDiagnostics(Constants.Encryption.DiagnosticsDecryptOperation);
         }
         response.result.forEach((operationResponse: OperationResponse, index: number) => {
           orderedResponses[batch.indexes[index]] = operationResponse;
         });
       } catch (err: any) {
+        if (this.clientContext.enableEncryption) {
+          await this.container.throwIfRequestNeedsARetryPostPolicyRefresh(err);
+        }
         // In the case of 410 errors, we need to recompute the partition key ranges
         // and redo the batch request, however, 410 errors occur for unsupported
         // partition key types as well since we don't support them, so for now we throw
@@ -853,7 +877,10 @@ export class Items {
       if (this.clientContext.enableEncryption) {
         operations = copyObject(operations);
         options = options || {};
-        addContainerRid(this.container);
+        if (!this.container.isEncryptionInitialized) {
+          await this.container.initializeEncryption();
+        }
+        this.container.encryptionProcessor.containerRid = this.container._rid;
         options.containerRid = this.container._rid;
         if (partitionKey) {
           const partitionKeyInternal = convertToInternalPartitionKey(partitionKey);
@@ -862,7 +889,7 @@ export class Items {
               partitionKeyInternal,
             );
         }
-        operations = await this.bulkBatchEncryptionHelper(operations);
+        operations = await this.bulkBatchEncryptionHelper(operations, diagnosticNode);
       }
 
       try {
@@ -876,6 +903,9 @@ export class Items {
         });
 
         if (this.clientContext.enableEncryption) {
+          diagnosticNode.beginEncryptionDiagnostics(
+            Constants.Encryption.DiagnosticsDecryptOperation,
+          );
           for (const result of response.result) {
             if (result.resourceBody) {
               result.resourceBody = await this.container.encryptionProcessor.decrypt(
@@ -883,19 +913,24 @@ export class Items {
               );
             }
           }
+          diagnosticNode.endEncryptionDiagnostics(Constants.Encryption.DiagnosticsDecryptOperation);
         }
 
         return response;
       } catch (err: any) {
         if (this.clientContext.enableEncryption) {
-          await this.container.ThrowIfRequestNeedsARetryPostPolicyRefresh(err);
+          await this.container.throwIfRequestNeedsARetryPostPolicyRefresh(err);
         }
         throw new Error(`Batch request error: ${err.message}`);
       }
     }, this.clientContext);
   }
 
-  private async bulkBatchEncryptionHelper(operations: OperationInput[]): Promise<OperationInput[]> {
+  private async bulkBatchEncryptionHelper(
+    operations: OperationInput[],
+    diagnosticNode: DiagnosticNodeInternal,
+  ): Promise<OperationInput[]> {
+    diagnosticNode.beginEncryptionDiagnostics(Constants.Encryption.DiagnosticsEncryptOperation);
     for (const operation of operations) {
       if (Object.prototype.hasOwnProperty.call(operation, "partitionKey")) {
         const partitionKeyInternal = convertToInternalPartitionKey(operation.partitionKey);
@@ -936,6 +971,7 @@ export class Items {
           break;
       }
     }
+    diagnosticNode.endEncryptionDiagnostics(Constants.Encryption.DiagnosticsEncryptOperation);
     return operations;
   }
 }
