+++ sdk/cosmosdb/cosmos/src/common/helper.ts
@@ -1,7 +1,6 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 import { CosmosClientOptions } from "../CosmosClientOptions";
-import { Container } from "../client";
 import {
   Serializer,
   NumberSerializer,
@@ -444,12 +443,5 @@ export function createSerializer(
       default:
         throw new Error("Invalid or Unsupported data type passed.");
     }
-  }
-}
-
-export async function addContainerRid(container: Container): Promise<void> {
-  if (!container._rid) {
-    const { resource: containerDefinition } = await this.container.read();
-    this.container._rid = containerDefinition._rid;
   }
 }
