+++ sdk/cosmosdb/cosmos/src/diagnostics/CosmosDiagnosticsContext.ts
@@ -1,15 +1,16 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+import { Constants } from "../common";
 import type {
   ClientSideRequestStatistics,
+  EncryptionDiagnostics,
   FailedRequestAttemptDiagnostic,
   GatewayStatistics,
   MetadataLookUpDiagnostic,
   MetadataLookUpType,
 } from "../CosmosDiagnostics";
 import { getCurrentTimestampInMs } from "../utils/time";
-
 /**
  * @hidden
  * Internal class to hold CosmosDiagnostic aggregate information all through the lifecycle of a request.
@@ -25,6 +26,7 @@ export class CosmosDiagnosticContext {
   private metadataLookups: MetadataLookUpDiagnostic[] = [];
   private gaterwayStatistics: GatewayStatistics[] = [];
   public locationEndpointsContacted: Set<string> = new Set();
+  encryptionDiagnostics: EncryptionDiagnostics;
 
   public constructor() {
     this.requestStartTimeUTCinMs = getCurrentTimestampInMs();
@@ -51,6 +53,15 @@ export class CosmosDiagnosticContext {
 
   public recordNetworkCall(gaterwayStatistics: GatewayStatistics): void {
     this.gaterwayStatistics.push(gaterwayStatistics);
+  }
+
+  public recordEncryptionDiagnostics(encryptionDiagnostics: EncryptionDiagnostics): void {
+    const { encryptContent, decryptContent } = encryptionDiagnostics;
+    const encryptionDuration = encryptContent[Constants.Encryption.DiagnosticsDuration] ?? 0;
+    const decryptionDuration = decryptContent[Constants.Encryption.DiagnosticsDuration] ?? 0;
+
+    encryptionDiagnostics.processingDurationInMs = encryptionDuration + decryptionDuration;
+    this.encryptionDiagnostics = encryptionDiagnostics;
   }
 
   /**
@@ -103,6 +114,7 @@ export class CosmosDiagnosticContext {
         failedAttempts: [...this.failedAttempts],
       },
       gatewayStatistics: this.gaterwayStatistics,
+      encryptionDiagnostics: this.encryptionDiagnostics,
     };
   }
