+++ sdk/cosmosdb/cosmos/src/diagnostics/DiagnosticNodeInternal.ts
@@ -2,9 +2,10 @@
 // Licensed under the MIT License.
 
 import { CosmosDiagnosticContext } from "./CosmosDiagnosticsContext";
-import type { RequestContext } from "../request";
-import type {
+import { ErrorResponse, RequestContext } from "../request";
+import {
   DiagnosticNode,
+  EncryptionDiagnostics,
   MetadataLookUpType,
   ClientConfigDiagnostic,
 } from "../CosmosDiagnostics";
@@ -34,6 +35,7 @@ export class DiagnosticNodeInternal implements DiagnosticNode {
   public durationInMs: number;
   public diagnosticLevel: CosmosDbDiagnosticLevel;
   private diagnosticCtx: CosmosDiagnosticContext;
+  private encryptionDiagnostics: EncryptionDiagnostics;
 
   /**
    * @internal
@@ -55,6 +57,13 @@ export class DiagnosticNodeInternal implements DiagnosticNode {
     this.parent = parent;
     this.diagnosticCtx = ctx;
     this.diagnosticLevel = diagnosticLevel;
+
+    // Initialize EncryptionDiagnostics
+    this.encryptionDiagnostics = {
+      encryptContent: {},
+      decryptContent: {},
+      processingDurationInMs: 0,
+    };
   }
 
   /**
@@ -256,6 +265,63 @@ export class DiagnosticNodeInternal implements DiagnosticNode {
         this.data.queryRecordsRead = previousCount + resources.length;
       }
     }
+  }
+
+  /**
+   * @internal
+   */
+  public beginEncryptionDiagnostics(operation: string): void {
+    const startTime = getCurrentTimestampInMs();
+    switch (operation) {
+      case Constants.Encryption.DiagnosticsEncryptOperation:
+        this.encryptionDiagnostics.encryptContent[Constants.Encryption.DiagnosticsStartTime] =
+          startTime;
+        break;
+      case Constants.Encryption.DiagnosticsDecryptOperation:
+        this.encryptionDiagnostics.decryptContent[Constants.Encryption.DiagnosticsStartTime] =
+          startTime;
+        break;
+      default:
+        throw new ErrorResponse("Invalid operation type for encryption diagnostics");
+    }
+  }
+  /**
+   * @internal
+   */
+  public endEncryptionDiagnostics(operation: string, propertiesCount?: number): void {
+    const endTime = getCurrentTimestampInMs();
+    let processingDuration = 0;
+    switch (operation) {
+      case Constants.Encryption.DiagnosticsEncryptOperation:
+        processingDuration =
+          endTime -
+          this.encryptionDiagnostics.encryptContent[Constants.Encryption.DiagnosticsStartTime];
+        this.encryptionDiagnostics.encryptContent[Constants.Encryption.DiagnosticsDuration] =
+          processingDuration;
+        // will be undefined in case of bulk/batch
+        if (propertiesCount !== undefined) {
+          this.encryptionDiagnostics.encryptContent[
+            Constants.Encryption.DiagnosticsPropertiesEncryptedCount
+          ] = propertiesCount;
+        }
+        break;
+
+      case Constants.Encryption.DiagnosticsDecryptOperation:
+        processingDuration =
+          endTime -
+          this.encryptionDiagnostics.decryptContent[Constants.Encryption.DiagnosticsStartTime];
+        this.encryptionDiagnostics.decryptContent[Constants.Encryption.DiagnosticsDuration] =
+          processingDuration;
+        if (propertiesCount !== undefined) {
+          this.encryptionDiagnostics.decryptContent[
+            Constants.Encryption.DiagnosticsPropertiesDecryptedCount
+          ] = propertiesCount;
+        }
+        break;
+      default:
+        throw new ErrorResponse("Invalid operation type for encryption diagnostics");
+    }
+    this.diagnosticCtx.recordEncryptionDiagnostics(this.encryptionDiagnostics);
   }
 
   /**
