+++ sdk/cosmosdb/cosmos/src/encryption/EncryptionProcessor.ts
@@ -4,7 +4,7 @@
 import { EncryptionSettings } from "./EncryptionSettings";
 import { EncryptionSettingForProperty } from "./EncryptionSettingForProperty";
 import { AeadAes256CbcHmacSha256Algorithm } from "./AeadAes256CbcHmacSha256Algorithm";
-import { ContainerDefinition, DatabaseDefinition, ItemDefinition } from "../client";
+import { ContainerDefinition, Database, ItemDefinition } from "../client";
 import { PartitionKeyInternal } from "../documents";
 import { TypeMarker } from "./enums/TypeMarker";
 import { ClientContext } from "../ClientContext";
@@ -25,15 +25,23 @@ import { EncryptionManager } from "./EncryptionManager";
 export class EncryptionProcessor {
   constructor(
     private readonly containerId: string,
-    private readonly databaseId: string,
+    public containerRid: string,
+    private readonly database: Database,
     private readonly clientContext: ClientContext,
     private encryptionManager: EncryptionManager,
   ) {}
 
-  async encrypt<T extends ItemDefinition>(body: T): Promise<T> {
+  async encrypt<T extends ItemDefinition>(
+    body: T,
+    diagnosticNode?: DiagnosticNodeInternal,
+  ): Promise<T> {
     if (!body) {
       throw new ErrorResponse("Input body is null or undefined.");
     }
+    if (diagnosticNode !== undefined) {
+      diagnosticNode.beginEncryptionDiagnostics(Constants.Encryption.DiagnosticsEncryptOperation);
+    }
+    let propertiesEncryptedCount = 0;
     const encryptionSettings = await this.getEncryptionSetting();
     if (!encryptionSettings) return body;
     for (const pathToEncrypt of encryptionSettings.pathsToEncrypt) {
@@ -51,8 +59,23 @@ export class EncryptionProcessor {
         settingForProperty,
         propertyName === "id",
       );
+      propertiesEncryptedCount++;
+    }
+    if (diagnosticNode !== undefined) {
+      diagnosticNode.endEncryptionDiagnostics(
+        Constants.Encryption.DiagnosticsEncryptOperation,
+        propertiesEncryptedCount,
+      );
     }
     return body;
+  }
+
+  async isPathEncrypted(path: string): Promise<boolean> {
+    path = extractPath(path);
+    const encryptionSettings = await this.getEncryptionSetting();
+    const settingForProperty = encryptionSettings.getEncryptionSettingForProperty(path);
+    if (!settingForProperty) return false;
+    return true;
   }
 
   async encryptProperty(path: string, value: any): Promise<any> {
@@ -194,10 +217,17 @@ export class EncryptionProcessor {
     return encryptedValue;
   }
 
-  async decrypt<T extends ItemDefinition>(body: T): Promise<T> {
+  async decrypt<T extends ItemDefinition>(
+    body: T,
+    diagnosticNode?: DiagnosticNodeInternal,
+  ): Promise<T> {
     if (body == null) {
       return body;
     }
+    if (diagnosticNode !== undefined) {
+      diagnosticNode.beginEncryptionDiagnostics(Constants.Encryption.DiagnosticsDecryptOperation);
+    }
+    let propertiesDecryptedCount = 0;
     const encryptionSettings = await this.getEncryptionSetting();
     if (!encryptionSettings) return body;
     for (const pathToEncrypt of encryptionSettings.pathsToEncrypt) {
@@ -214,6 +244,13 @@ export class EncryptionProcessor {
         body[propertyName],
         settingForProperty,
         propertyName === "id",
+      );
+      propertiesDecryptedCount++;
+    }
+    if (diagnosticNode !== undefined) {
+      diagnosticNode.endEncryptionDiagnostics(
+        Constants.Encryption.DiagnosticsDecryptOperation,
+        propertiesDecryptedCount,
       );
     }
     return body;
@@ -275,13 +312,13 @@ export class EncryptionProcessor {
   }
 
   async getEncryptionSetting(forceRefresh?: boolean): Promise<EncryptionSettings> {
-    const key = this.databaseId + "/" + this.containerId;
+    const key = this.database._rid + "/" + this.containerRid;
     const encryptionSetting =
       this.encryptionManager.encryptionSettingsCache.getEncryptionSettings(key);
     if (forceRefresh || !encryptionSetting) {
       return withDiagnostics(async (diagnosticNode: DiagnosticNodeInternal) => {
-        const path = `/dbs/${this.databaseId}/colls/${this.containerId}`;
-        const id = `dbs/${this.databaseId}/colls/${this.containerId}`;
+        const path = `/dbs/${this.database.id}/colls/${this.containerId}`;
+        const id = `dbs/${this.database.id}/colls/${this.containerId}`;
         const response = await this.clientContext.read<ContainerDefinition>({
           path,
           resourceType: ResourceType.container,
@@ -303,11 +340,11 @@ export class EncryptionProcessor {
     }
     return encryptionSetting;
   }
+
   private async buildEncryptionAlgorithm(
     propertySetting: EncryptionSettingForProperty,
   ): Promise<AeadAes256CbcHmacSha256Algorithm> {
-    const key = `${this.databaseId}/${propertySetting.encryptionKeyId}`;
-
+    const key = `${this.database._rid}/${propertySetting.encryptionKeyId}`;
     let clientEncryptionKeyProperties =
       this.encryptionManager.clientEncryptionKeyPropertiesCache.getClientEncryptionKeyProperties(
         key,
@@ -360,8 +397,8 @@ export class EncryptionProcessor {
     cekEtag?: string,
   ): Promise<ClientEncryptionKeyProperties> {
     return withDiagnostics(async (diagnosticNode: DiagnosticNodeInternal) => {
-      const path = `/dbs/${this.databaseId}/clientencryptionkeys/${cekId}`;
-      const id = `dbs/${this.databaseId}/clientencryptionkeys/${cekId}`;
+      const path = `/dbs/${this.database.id}/clientencryptionkeys/${cekId}`;
+      const id = `dbs/${this.database.id}/clientencryptionkeys/${cekId}`;
       const options: RequestOptions = {};
       if (cekEtag) {
         options.accessCondition = {
@@ -369,15 +406,7 @@ export class EncryptionProcessor {
           condition: cekEtag,
         };
       }
-      const dbResponse = await this.clientContext.read<DatabaseDefinition>({
-        path: `/dbs/${this.databaseId}`,
-        resourceType: ResourceType.database,
-        resourceId: `dbs/${this.databaseId}`,
-        options,
-        diagnosticNode,
-      });
-      const dbRid = dbResponse.result._rid;
-      options.databaseRid = dbRid;
+      options.databaseRid = this.database._rid;
       const response = await this.clientContext.read<ClientEncryptionKeyRequest>({
         path: path,
         resourceType: ResourceType.clientencryptionkey,
@@ -387,7 +416,7 @@ export class EncryptionProcessor {
       });
       if (response.code === StatusCodes.NotModified) {
         throw new ErrorResponse(
-          `The Client Encryption Key with key id: ${cekId} on database: ${this.databaseId} needs to be rewrapped with a valid Key Encryption Key using rewrapClientEncryptionKey. The Key Encryption Key used to wrap the Client Encryption Key has been revoked`,
+          `The Client Encryption Key with key id: ${cekId} on database: ${this.database.id} needs to be rewrapped with a valid Key Encryption Key using rewrapClientEncryptionKey. The Key Encryption Key used to wrap the Client Encryption Key has been revoked`,
         );
       }
       const clientEncryptionKeyProperties = new ClientEncryptionKeyProperties(
@@ -397,7 +426,7 @@ export class EncryptionProcessor {
         Buffer.from(response.result.wrappedDataEncryptionKey, "base64"),
         response.result.keyWrapMetadata,
       );
-      const key = this.databaseId + "/" + cekId;
+      const key = this.database._rid + "/" + cekId;
       this.encryptionManager.clientEncryptionKeyPropertiesCache.setClientEncryptionKeyProperties(
         key,
         clientEncryptionKeyProperties,
