+++ sdk/cosmosdb/cosmos/src/encryption/index.ts
@@ -15,3 +15,4 @@ export { ProtectedDataEncryptionKey } from "./EncryptionKey";
 export { EncryptionProcessor } from "./EncryptionProcessor";
 export { EncryptionQueryBuilder } from "./EncryptionQueryBuilder";
 export { EncryptionManager } from "./EncryptionManager";
+export { EncryptionTimeToLive } from "./EncryptionTimeToLive";
