+++ sdk/cosmosdb/cosmos/src/index.ts
@@ -118,6 +118,7 @@ export {
   ClientSideRequestStatistics,
   ClientConfigDiagnostic,
   DiagnosticNode,
+  EncryptionDiagnostics,
 } from "./CosmosDiagnostics";
 
 export {
