+++ sdk/cosmosdb/cosmos/src/queryIterator.ts
@@ -4,7 +4,7 @@
 /// <reference lib="esnext.asynciterable" />
 import type { ClientContext } from "./ClientContext";
 import { DiagnosticNodeInternal, DiagnosticNodeType } from "./diagnostics/DiagnosticNodeInternal";
-import { addContainerRid, getPathFromLink, ResourceType, StatusCodes } from "./common";
+import { getPathFromLink, ResourceType, StatusCodes } from "./common";
 import {
   CosmosHeaders,
   ExecutionContext,
@@ -111,7 +111,7 @@ export class QueryIterator<T> {
         response = await this.queryExecutionContext.fetchMore(diagnosticNode);
       } catch (error: any) {
         if (this.container && this.clientContext.enableEncryption) {
-          await this.container.ThrowIfRequestNeedsARetryPostPolicyRefresh(error);
+          await this.container.throwIfRequestNeedsARetryPostPolicyRefresh(error);
         }
         if (this.needsQueryPlan(error)) {
           await this.createExecutionContext(diagnosticNode);
@@ -133,7 +133,7 @@ export class QueryIterator<T> {
         response.result.length
       ) {
         for (let item of response.result) {
-          item = await this.container.encryptionProcessor.decrypt(item);
+          item = await this.container.encryptionProcessor.decrypt(item, diagnosticNode);
         }
       }
 
@@ -174,7 +174,7 @@ export class QueryIterator<T> {
         response = await this.fetchAllInternal(diagnosticNode);
       } catch (error: any) {
         if (this.container && this.clientContext.enableEncryption) {
-          await this.container.ThrowIfRequestNeedsARetryPostPolicyRefresh(error);
+          await this.container.throwIfRequestNeedsARetryPostPolicyRefresh(error);
         }
         throw error;
       }
@@ -186,7 +186,7 @@ export class QueryIterator<T> {
         response.resources.length > 0
       ) {
         for (let result of response.resources) {
-          result = await this.container.encryptionProcessor.decrypt(result);
+          result = await this.container.encryptionProcessor.decrypt(result, diagnosticNode);
         }
       }
       return response;
@@ -231,7 +231,7 @@ export class QueryIterator<T> {
         response = await this.queryExecutionContext.fetchMore(diagnosticNode);
       } catch (error: any) {
         if (this.container && this.clientContext.enableEncryption) {
-          await this.container.ThrowIfRequestNeedsARetryPostPolicyRefresh(error);
+          await this.container.throwIfRequestNeedsARetryPostPolicyRefresh(error);
         }
         if (this.needsQueryPlan(error)) {
           await this.createExecutionContext(diagnosticNode);
@@ -252,7 +252,7 @@ export class QueryIterator<T> {
         response.result.length
       ) {
         for (let result of response.result) {
-          result = await this.container.encryptionProcessor.decrypt(result);
+          result = await this.container.encryptionProcessor.decrypt(result, diagnosticNode);
         }
       }
       return new FeedResponse<T>(
@@ -414,7 +414,10 @@ export class QueryIterator<T> {
   private async init(diagnosticNode: DiagnosticNodeInternal): Promise<void> {
     // add rid to options if encryption is enable for client
     if (this.container && this.clientContext.enableEncryption) {
-      addContainerRid(this.container);
+      if (!this.container.isEncryptionInitialized) {
+        await this.container.initializeEncryption();
+      }
+      this.container.encryptionProcessor.containerRid = this.container._rid;
       this.options.containerRid = this.container._rid;
     }
     if (this.isInitialized === true) {
