+++ sdk/cosmosdb/cosmos/test/public/functional/clientSideEncryption.spec.ts
@@ -1,720 +1,2152 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
+import { randomUUID } from "@azure/core-util";
 import {
   CosmosClient,
-  EncryptionAlgorithm,
-  EncryptionKeyResolverName,
   EncryptionKeyWrapMetadata,
-  KeyEncryptionKeyAlgorithm,
   Database,
+  Container,
+  ContainerDefinition,
+  ClientEncryptionPolicy,
+  KeyEncryptionKeyAlgorithm,
   ClientEncryptionIncludedPath,
   EncryptionType,
-  ClientEncryptionPolicy,
+  EncryptionAlgorithm,
+  EncryptionKeyResolverName,
   StatusCodes,
-  PatchOperation,
-  PartitionKeyDefinitionVersion,
-  PartitionKeyKind,
-  OperationInput,
-  PatchOperationType,
   BulkOperationType,
-  EncryptionKeyResolver,
+  ChangeFeedStartFrom,
   EncryptionQueryBuilder,
+  OperationInput,
+  PatchOperation,
+  PatchOperationType,
+  PartitionKeyDefinitionVersion,
+  ChangeFeedMode,
+  ChangeFeedPolicy,
+  ChangeFeedRetentionTimeSpan,
+  PartitionKeyKind,
+  PermissionMode,
 } from "../../../src";
-import { assert } from "chai";
 import { masterKey } from "../common/_fakeTestSecrets";
 import { endpoint } from "../common/_testConfig";
-
+import {
+  compareMetadata,
+  MockKeyVaultEncryptionKeyResolver,
+  testCreateClientEncryptionKey,
+  testCreateItem,
+  TestDoc,
+  verifyExpectedDocResponse,
+  verifyItemByRead,
+  testRewrapClientEncryptionKey,
+  testUpsertItem,
+  validateQueryResults,
+  testPatchItem,
+  testReplaceItem,
+  testDeleteItem,
+  verifyDiagnostics,
+} from "../common/encryptionTestHelpers";
 import { removeAllDatabases } from "../common/TestHelpers";
-import { randomUUID } from "@azure/core-util";
+import { assert } from "chai";
+import { EncryptionTimeToLive } from "../../../src/encryption/EncryptionTimeToLive";
 
-export class MockKeyVaultEncryptionKeyResolver implements EncryptionKeyResolver {
-  private keyInfo: { [key: string]: number } = {
-    tempmetadata1: 1,
-    tempmetadata2: 2,
-    "revokedKek-metadata": 3,
-    mymetadata1: 4,
-    mymetadata2: 5,
-  };
-  revokeAccessSet = false;
-  wrapKeyCallsCount: { [key: string]: number };
-  unwrapKeyCallsCount: { [key: string]: number };
-  constructor() {
-    this.wrapKeyCallsCount = {};
-    this.unwrapKeyCallsCount = {};
-    this.revokeAccessSet = false;
-  }
-  async unwrapKey(encryptionKeyId: string, algorithm: string, key: Buffer): Promise<Buffer> {
-    console.log(algorithm);
-    if (encryptionKeyId === "revokedKek-metadata" && this.revokeAccessSet) {
-      throw new Error("Forbidden");
-    }
-    console.log("here");
-    if (!this.unwrapKeyCallsCount.encryptedKeyId) {
-      this.unwrapKeyCallsCount[encryptionKeyId] = 1;
-    } else {
-      this.unwrapKeyCallsCount.encryptedKeyId++;
-    }
-    const moveBy = this.keyInfo[encryptionKeyId];
-    const plainKey = Buffer.alloc(key.length);
-    for (let i = 0; i < key.length; i++) {
-      plainKey[i] = key[i] - moveBy;
-    }
-    return plainKey;
-  }
+let encryptionClient: CosmosClient;
+let metadata1: EncryptionKeyWrapMetadata;
+let metadata2: EncryptionKeyWrapMetadata;
+let database: Database;
+let encryptionContainer: Container;
+let encryptionContainerForChangeFeed: Container;
+let testKeyEncryptionKeyResolver: MockKeyVaultEncryptionKeyResolver;
+let containerDefinition: ContainerDefinition;
+let clientEncryptionPolicy: ClientEncryptionPolicy;
 
-  async wrapKey(encryptionKeyId: string, algorithm: string, wrappedKey: Buffer): Promise<Buffer> {
-    console.log(algorithm);
-    if (!this.wrapKeyCallsCount.encryptedKeyId) {
-      this.wrapKeyCallsCount[encryptionKeyId] = 1;
-    } else {
-      this.wrapKeyCallsCount.encryptedKeyId++;
-    }
-    const moveBy = this.keyInfo[encryptionKeyId];
-    const encryptedKey = Buffer.alloc(wrappedKey.length);
-    for (let i = 0; i < wrappedKey.length; i++) {
-      encryptedKey[i] = wrappedKey[i] + moveBy;
-    }
-    return encryptedKey;
-  }
-}
 const testKeyVault = "TESTKEYSTORE_VAULT" as EncryptionKeyResolverName;
 
-const client = new CosmosClient({
-  endpoint: endpoint,
-  key: masterKey,
-  enableEncryption: true,
-  keyEncryptionKeyResolver: new MockKeyVaultEncryptionKeyResolver(),
-  encryptionKeyResolverName: testKeyVault,
-});
-
 describe("Client Side Encryption", () => {
-  let database: Database;
   before(async () => {
     await removeAllDatabases();
-    try {
-      const response = await client.databases.createIfNotExists({ id: "encryptionTestDatabase2" });
-      database = response.database;
-      console.log("Database created");
-      await database.createClientEncryptionKey(
-        "dek1",
-        EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
-        new EncryptionKeyWrapMetadata(
-          testKeyVault,
-          "key1",
-          "tempmetadata1",
-          KeyEncryptionKeyAlgorithm.RSA_OAEP,
-        ),
-      );
-    } catch (error) {
-      console.error("Error creating database or encryption key:", error);
-    }
-  });
-
-  it("CRUD operations on encryption-enabled container with non-encrypted partition key/id", async () => {
-    const includedPaths = [
-      "/valueint",
-      "/valuedouble",
-      "/valuestring",
-      "/valuebit",
-      "/object",
-      "/array",
-    ].map(
-      (path) =>
-        new ClientEncryptionIncludedPath(
-          path,
-          "dek1",
-          EncryptionType.DETERMINISTIC,
-          EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
-        ),
+    testKeyEncryptionKeyResolver = new MockKeyVaultEncryptionKeyResolver();
+    metadata1 = new EncryptionKeyWrapMetadata(
+      testKeyVault,
+      "key1",
+      "tempmetadata1",
+      KeyEncryptionKeyAlgorithm.RSA_OAEP,
     );
-    const clientEncryptionPolicy = new ClientEncryptionPolicy(includedPaths);
-    const { container } = await database.containers.createIfNotExists({
-      id: "encryptionContainer",
-      partitionKey: "/id",
-      clientEncryptionPolicy: clientEncryptionPolicy,
-    });
-    await container.initializeEncryption();
-
-    // create Item
-    const createResponse = await container.items.create({
-      id: "1",
-      valueint: 32,
-      valuedouble: 3.14,
-      valuestring: "hello",
-      valuebit: true,
-      object: {
-        o1: 32,
-        o2: 3.14,
-      },
-      array: ["hello", true],
-    });
-    assert.equal(createResponse.statusCode, StatusCodes.Created);
-    // upsert Item
-    const upsertResponse = await container.items.upsert({
-      id: "1",
-      valueint: 105,
-      valuedouble: 8.5,
-      valuestring: "hello",
-      valuebit: true,
-      object: {
-        o1: 32,
-        o2: 3.14,
-      },
-      array: ["hello", true],
-    });
-    assert.equal(upsertResponse.statusCode, StatusCodes.Ok);
-
-    // read item
-    const readResponse = await container.item("1", "1").read();
-    assert.equal(readResponse.statusCode, StatusCodes.Ok);
-
-    // replace item
-    const replaceResponse = await container.item("1", "1").replace({
-      id: "1",
-      valueint: 500,
-      valuedouble: 8.5,
-      valuestring: "hello",
-      valuebit: true,
-      object: {
-        o1: 32,
-        o2: 3.14,
-      },
-      array: ["hello", true],
-    });
-    assert.equal(replaceResponse.statusCode, StatusCodes.Ok);
-
-    // patch item
-    const patchOperations: PatchOperation[] = [
-      { op: "replace", path: "/valueint", value: 1000 },
-      { op: "remove", path: "/valuebit" },
-    ];
-    const patchResponse = await container.item("1", "1").patch(patchOperations);
-    assert.equal(patchResponse.statusCode, StatusCodes.Ok);
-
-    // delete operation
-    const deleteResponse = await container.item("1", "1").delete();
-    assert.equal(deleteResponse.statusCode, StatusCodes.NoContent);
-  });
-
-  it("should validate policy format version", async () => {
-    const path = new ClientEncryptionIncludedPath(
-      "/id",
-      "dek1",
-      EncryptionType.DETERMINISTIC,
-      EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
+    metadata2 = new EncryptionKeyWrapMetadata(
+      testKeyVault,
+      "key2",
+      "tempmetadata2",
+      KeyEncryptionKeyAlgorithm.RSA_OAEP,
     );
-    const clientEncryptionPolicy = new ClientEncryptionPolicy([path]);
-    const { container } = await database.containers.createIfNotExists({
-      id: "encryptionContainerWithEncryptedId",
-      partitionKey: "/id",
-      clientEncryptionPolicy: clientEncryptionPolicy,
-    });
-    try {
-      await container.initializeEncryption();
-      assert.fail("Expected an error to be thrown");
-    } catch (error) {
-      const errorMessage = error.message;
-      const exepectedErrorMessage =
-        "Encryption of partition key or id is only supported with policy format version 2.";
-      assert.equal(errorMessage, exepectedErrorMessage);
-    }
-  });
-
-  it("should validate encryption type with policy version 2.", async () => {
-    const path = new ClientEncryptionIncludedPath(
-      "/id",
-      "dek1",
-      EncryptionType.RANDOMIZED,
-      EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
-    );
-    const clientEncryptionPolicy = new ClientEncryptionPolicy([path], 2);
-    const { container } = await database.containers.createIfNotExists({
-      id: "encryptionContainerWithEncryptedId1",
-      partitionKey: "/id",
-      clientEncryptionPolicy: clientEncryptionPolicy,
-    });
-    try {
-      await container.initializeEncryption();
-      assert.fail("Expected an error to be thrown");
-    } catch (error) {
-      const errorMessage = error.message;
-      const exepectedErrorMessage =
-        "Encryption Type must be deterministic for encryption of partition key/id";
-      assert.equal(errorMessage, exepectedErrorMessage);
-    }
-  });
-
-  it("CRUD operation with encrypted hierarchical partition key", async () => {
-    const includedPaths = ["/valueint", "/valuebit", "/object", "/id"].map(
-      (path) =>
-        new ClientEncryptionIncludedPath(
-          path,
-          "dek1",
-          EncryptionType.DETERMINISTIC,
-          EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
-        ),
-    );
-    const clientEncryptionPolicy = new ClientEncryptionPolicy(includedPaths, 2);
-
-    const containerDefinition = {
-      id: "hierarchicalEncryptionContainer",
-      partitionKey: {
-        paths: ["/valuestring", "/object/o1"],
-        version: PartitionKeyDefinitionVersion.V2,
-        kind: PartitionKeyKind.MultiHash,
-      },
-      clientEncryptionPolicy: clientEncryptionPolicy,
-    };
-    const { container } = await database.containers.createIfNotExists(containerDefinition);
-    // console.log(container.id);
-    await container.initializeEncryption();
-
-    // create Item
-    const createResponse = await container.items.create({
-      id: "1",
-      valueint: 32,
-      valuedouble: 3.14,
-      valuestring: "hello",
-      valuebit: true,
-      object: {
-        o1: 32,
-        o2: 3.14,
-      },
-      array: ["hello", true],
-    });
-
-    assert.equal(createResponse.statusCode, StatusCodes.Created);
-
-    // upsert Item
-    const upsertResponse = await container.items.upsert({
-      id: "1",
-      valueint: 100,
-      valuedouble: 3.14,
-      valuestring: "hello",
-      valuebit: true,
-      object: {
-        o1: 32,
-        o2: 3.14,
-      },
-      array: ["hello", true],
-    });
-
-    assert.equal(upsertResponse.statusCode, StatusCodes.Ok);
-
-    const upsertResponse1 = await container.items.upsert({
-      id: "1",
-      valueint: 100,
-      valuedouble: 3.14,
-      valuestring: "hello",
-      valuebit: true,
-      object: {
-        o1: 35,
-        o2: 3.14,
-      },
-      array: ["hello", true],
-    });
-    assert.equal(upsertResponse1.statusCode, StatusCodes.Created);
-
-    // read item
-    const readResponse = await container.item("1", ["hello", 32]).read();
-    assert.equal(readResponse.statusCode, StatusCodes.Ok);
-    // replace item
-    const replaceResponse = await container.item("1", ["hello", 32]).replace({
-      id: "1",
-      valueint: 32,
-      valuedouble: 8.5,
-      valuestring: "hello",
-      valuebit: false,
-      object: {
-        o1: 32,
-        o2: 3.14,
-      },
-      array: ["hello", true],
-    });
-    assert.equal(replaceResponse.statusCode, StatusCodes.Ok);
-    // patch item
-    const patchOperations: PatchOperation[] = [
-      { op: "add", path: "/valuebit", value: true },
-      { op: "replace", path: "/valueint", value: 1000 },
-      { op: "remove", path: "/valuebit" },
-    ];
-    const patchResponse = await container.item("1", ["hello", 32]).patch(patchOperations);
-    assert.equal(patchResponse.statusCode, StatusCodes.Ok);
-    // delete operation
-    const deleteResponse = await container.item("1", ["hello", 32]).delete();
-    assert.equal(deleteResponse.statusCode, StatusCodes.NoContent);
-  });
-
-  it("encryption batch operation", async () => {
-    const includedPaths = ["/key", "/id", "/object"].map(
-      (path) =>
-        new ClientEncryptionIncludedPath(
-          path,
-          "dek1",
-          EncryptionType.DETERMINISTIC,
-          EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
-        ),
-    );
-    const clientEncryptionPolicy = new ClientEncryptionPolicy(includedPaths, 2);
-    const containerDef = {
-      id: "encryption_container_batch",
-      partitionKey: "/key",
-      clientEncryptionPolicy: clientEncryptionPolicy,
-    };
-    const { container } = await database.containers.createIfNotExists(containerDef);
-    await container.initializeEncryption();
-    const operations: OperationInput[] = [
-      {
-        operationType: "Create",
-        resourceBody: { id: "doc1", name: "sample", key: "A", object: { o1: 1, o2: 2 } },
-      },
-      {
-        operationType: "Upsert",
-        resourceBody: { id: "doc2", name: "other", key: "A", object: { o1: 1, o2: 2 } },
-      },
-      {
-        operationType: "Read",
-        id: "doc1",
-      },
-      {
-        operationType: "Delete",
-        id: "doc1",
-      },
-      {
-        operationType: "Patch",
-        id: "doc2",
-        resourceBody: {
-          operations: [{ op: PatchOperationType.set, path: "/name", value: "sample" }],
-          condition: "from c where NOT IS_DEFINED(c.newImproved)",
-        },
-      },
-    ];
-
-    const response = await container.items.batch(operations, "A");
-    assert.equal(response.result[0].statusCode, StatusCodes.Created);
-    assert.equal(response.result[1].statusCode, StatusCodes.Created);
-    assert.equal(response.result[2].statusCode, StatusCodes.Ok);
-    assert.equal(response.result[3].statusCode, StatusCodes.NoContent);
-    assert.equal(response.result[4].statusCode, StatusCodes.Ok);
-  });
-
-  it("encryption batch with hierarchical partition key", async () => {
-    const includedPaths = ["/key", "/id", "/object"].map(
-      (path) =>
-        new ClientEncryptionIncludedPath(
-          path,
-          "dek1",
-          EncryptionType.DETERMINISTIC,
-          EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
-        ),
-    );
-    const clientEncryptionPolicy = new ClientEncryptionPolicy(includedPaths, 2);
-    const containerDef = {
-      id: "encryption_hierarchical_container_batch",
-      partitionKey: {
-        paths: ["/key", "/object/o1"],
-        version: PartitionKeyDefinitionVersion.V2,
-        kind: PartitionKeyKind.MultiHash,
-      },
-      clientEncryptionPolicy: clientEncryptionPolicy,
-    };
-    const { container } = await database.containers.createIfNotExists(containerDef);
-    await container.initializeEncryption();
-    const operations: OperationInput[] = [
-      {
-        operationType: "Create",
-        resourceBody: { id: "doc1", name: "sample", key: "A", object: { o1: 1, o2: 2 } },
-      },
-      {
-        operationType: "Upsert",
-        resourceBody: { id: "doc2", name: "other", key: "A", object: { o1: 1, o2: 2 } },
-      },
-      {
-        operationType: "Read",
-        id: "doc1",
-      },
-      {
-        operationType: "Delete",
-        id: "doc1",
-      },
-      {
-        operationType: "Patch",
-        id: "doc2",
-        resourceBody: {
-          operations: [{ op: PatchOperationType.set, path: "/name", value: "sample" }],
-          condition: "from c where NOT IS_DEFINED(c.newImproved)",
-        },
-      },
-    ];
-
-    const response = await container.items.batch(operations, ["A", 1]);
-    assert.equal(response.result[0].statusCode, StatusCodes.Created);
-    assert.equal(response.result[1].statusCode, StatusCodes.Created);
-    assert.equal(response.result[2].statusCode, StatusCodes.Ok);
-    assert.equal(response.result[3].statusCode, StatusCodes.NoContent);
-    assert.equal(response.result[4].statusCode, StatusCodes.Ok);
-  });
-
-  it("encryption bulk operation", async () => {
-    const includedPaths = ["/key", "/id"].map(
-      (path) =>
-        new ClientEncryptionIncludedPath(
-          path,
-          "dek1",
-          EncryptionType.DETERMINISTIC,
-          EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
-        ),
-    );
-    const clientEncryptionPolicy = new ClientEncryptionPolicy(includedPaths, 2);
-    const containerDef = {
-      id: "encryption_container_bulk",
-      partitionKey: "/key",
-      clientEncryptionPolicy: clientEncryptionPolicy,
-    };
-    const { container } = await database.containers.createIfNotExists(containerDef);
-    await container.initializeEncryption();
-    const operations = [
-      {
-        operationType: BulkOperationType.Create,
-        resourceBody: { id: "doc1", name: "sample", key: "A" },
-      },
-      {
-        operationType: BulkOperationType.Upsert,
-        partitionKey: "A",
-        resourceBody: { id: "doc2", name: "other", key: "A" },
-      },
-      {
-        operationType: BulkOperationType.Read,
-        id: "doc1",
-        partitionKey: "A",
-      },
-      {
-        operationType: BulkOperationType.Delete,
-        id: "doc2",
-        partitionKey: "A",
-      },
-      {
-        operationType: BulkOperationType.Replace,
-        partitionKey: "A",
-        id: "doc1",
-        resourceBody: { id: "doc2", name: "nice", key: "A" },
-      },
-    ];
-    const response = await container.items.bulk(operations);
-    assert.equal(response[0].statusCode, StatusCodes.Created);
-    assert.equal(response[1].statusCode, StatusCodes.Created);
-    assert.equal(response[2].statusCode, StatusCodes.Ok);
-    assert.equal(response[3].statusCode, StatusCodes.NoContent);
-    assert.equal(response[4].statusCode, StatusCodes.Ok);
-  });
-
-  it("encryption bulk operation with hierarchical partition key", async () => {
-    const includedPaths = ["/key", "/id"].map(
-      (path) =>
-        new ClientEncryptionIncludedPath(
-          path,
-          "dek1",
-          EncryptionType.DETERMINISTIC,
-          EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
-        ),
-    );
-    const clientEncryptionPolicy = new ClientEncryptionPolicy(includedPaths, 2);
-    const containerDef = {
-      id: "encryption_hierarchical_container_bulk",
-      partitionKey: {
-        paths: ["/key", "/object/o1"],
-        version: PartitionKeyDefinitionVersion.V2,
-        kind: PartitionKeyKind.MultiHash,
-      },
-      clientEncryptionPolicy: clientEncryptionPolicy,
-    };
-    const { container } = await database.containers.createIfNotExists(containerDef);
-    await container.initializeEncryption();
-    const operations = [
-      {
-        operationType: BulkOperationType.Create,
-        resourceBody: { id: "doc1", name: "sample", key: "A", object: { o1: 1, o2: 2 } },
-      },
-      {
-        operationType: BulkOperationType.Upsert,
-        partitionKey: ["B", 1],
-        resourceBody: { id: "doc1", name: "other", key: "B", object: { o1: 1, o2: 2 } },
-      },
-      {
-        operationType: BulkOperationType.Read,
-        id: "doc1",
-        partitionKey: ["A", 1],
-      },
-      {
-        operationType: BulkOperationType.Delete,
-        id: "doc1",
-        partitionKey: ["A", 1],
-      },
-      {
-        operationType: BulkOperationType.Replace,
-        partitionKey: ["B", 1],
-        id: "doc1",
-        resourceBody: { id: "doc1", name: "nice", key: "B", object: { o1: 1, o2: 2 } },
-      },
-    ];
-    const response = await container.items.bulk(operations);
-    assert.equal(response[0].statusCode, StatusCodes.Created);
-    assert.equal(response[1].statusCode, StatusCodes.Created);
-    assert.equal(response[2].statusCode, StatusCodes.Ok);
-    assert.equal(response[3].statusCode, StatusCodes.NoContent);
-    assert.equal(response[4].statusCode, StatusCodes.Ok);
-  });
-
-  it("encryption query iterator", async () => {
-    const includedPaths = [
-      "/key",
-      "/id",
-      "/boolval",
-      "/floatval",
-      "/arrayval",
-      "/objval",
-      "/date",
-    ].map(
-      (path) =>
-        new ClientEncryptionIncludedPath(
-          path,
-          "dek1",
-          EncryptionType.DETERMINISTIC,
-          EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
-        ),
-    );
-    const clientEncryptionPolicy = new ClientEncryptionPolicy(includedPaths, 2);
-    const containerDef = {
-      id: "encryption_query_container",
-      partitionKey: "/key",
-      clientEncryptionPolicy: clientEncryptionPolicy,
-    };
-    const { container } = await database.containers.createIfNotExists(containerDef);
-    await container.initializeEncryption();
-    // const dateString = "1987-12-25T00:00:00Z";
-    const date = new Date(1987, 12, 25);
-    const date2 = new Date(1987, 12, 25, 0, 0, 0);
-    const date3 = new Date(1987, 12, 25, 0, 0, 0, 500);
-
-    console.log(date);
-    console.log(date2);
-    console.log(date3);
-
-    console.log("CREATE ITEM");
-
-    await Promise.all([
-      // container.items.create({ id: "1", name: "foo", key: 1, boolval: true, floatval: 3.0 }),
-      // container.items.create({ id: "2", name: "bar", key: 2, boolval: false, floatval: 6.5 }),
-      // container.items.create({ id: "1", name: "foo", key: 3, boolval: true, floatval: 6.5 }),
-      // container.items.create({ id: "2", name: "bar", key: 4, boolval: false, floatval: 0.0 }),
-      container.items.create({
-        id: "5",
-        name: "foo",
-        key: 5,
-        boolval: true,
-        floatval: 6.5,
-        date: date,
-      }),
-      // container.items.create({
-      //   id: "2",
-      //   name: "bar",
-      //   key: 5,
-      //   boolval: false,
-      //   floatval: 0.0,
-      //   arrayval: [1, true, 3.4],
-      // }),
-      // container.items.create({
-      //   id: "3",
-      //   name: "bar",
-      //   key: 5,
-      //   boolval: false,
-      //   floatval: 0.0,
-      //   objval: { key: 1, boolval: true, floatval: 3.4 },
-      // }),
-    ]);
-    const readAllResponse = await container.items.readAll().fetchAll();
-
-    assert.equal(readAllResponse.resources.length, 1);
-
-    // const queryBuilder = new EncryptionQueryBuilder(
-    //   "SELECT * FROM f WHERE f.id = @id and f.boolval = @boolval",
-    // );
-    // queryBuilder.addStringParameter("@id", "1", "/id");
-    // queryBuilder.addBooleanParameter("@boolval", true, "/boolval");
-    // const iterator1 = await container.items.getEncryptionQueryIterator(queryBuilder);
-    // for await (const response of iterator1.getAsyncIterator()) {
-    //   assert.equal(response.resources.length, 2);
-    // }
-
-    // const queryBuilder1 = new EncryptionQueryBuilder(
-    //   "SELECT * FROM f WHERE f.id = @id and f.floatval = @floatval",
-    // );
-    // queryBuilder1.addStringParameter("@id", "1", "/id");
-    // queryBuilder1.addFloatParameter("@floatval", 6.5, "/floatval");
-
-    // const iterator2 = await container.items.getEncryptionQueryIterator(queryBuilder1);
-    // const fetchNextResponse = await iterator2.fetchNext();
-    // assert.equal(fetchNextResponse.resources.length, 1);
-
-    // const queryBuilder2 = new EncryptionQueryBuilder("SELECT * FROM f WHERE f.key = @key");
-    // queryBuilder2.addIntegerParameter("@key", 1, "/key");
-    // const iterator3 = await container.items.getEncryptionQueryIterator(queryBuilder2);
-    // const fetchAllResponse = await iterator3.fetchAll();
-    // assert.equal(fetchAllResponse.resources.length, 1);
-
-    // const queryBuilder3 = new EncryptionQueryBuilder(
-    //   "SELECT * FROM f WHERE f.arrayval = @arrayval",
-    // );
-    // queryBuilder3.addArrayParameter("@arrayval", [1, true, 3.4], "/arrayval");
-    // const iterator4 = await container.items.getEncryptionQueryIterator(queryBuilder3);
-    // const fetchAllResponse1 = await iterator4.fetchAll();
-    // assert.equal(fetchAllResponse1.resources.length, 1);
-
-    // const queryBuilder4 = new EncryptionQueryBuilder("SELECT * FROM f WHERE f.objval = @objval");
-    // queryBuilder4.addObjectParameter(
-    //   "@objval",
-    //   { key: 1, boolval: true, floatval: 3.4 },
-    //   "/objval",
-    // );
-    // const iterator5 = await container.items.getEncryptionQueryIterator(queryBuilder4);
-    // const fetchAllResponse2 = await iterator5.fetchAll();
-    // assert.equal(fetchAllResponse2.resources.length, 1);
-    console.log("QUERY ITEM");
-
-    const queryBuilder5 = new EncryptionQueryBuilder("SELECT * FROM f WHERE f.date = @date");
-    queryBuilder5.addDateParameter("@date", date, "/date");
-    const iterator6 = await container.items.getEncryptionQueryIterator(queryBuilder5);
-    const fetchAllResponse3 = await iterator6.fetchAll();
-    assert.equal(fetchAllResponse3.resources.length, 1);
-  });
-
-  it("kek revoke handling", async () => {
-    const keyEncryptionKeyResolver = new MockKeyVaultEncryptionKeyResolver();
-    const encryptionClient = new CosmosClient({
+    encryptionClient = new CosmosClient({
       endpoint: endpoint,
       key: masterKey,
       enableEncryption: true,
-      keyEncryptionKeyResolver: keyEncryptionKeyResolver,
-      encryptionKeyTimeToLiveInHours: 0,
+      keyEncryptionKeyResolver: testKeyEncryptionKeyResolver,
       encryptionKeyResolverName: testKeyVault,
+      encryptionKeyTimeToLive: EncryptionTimeToLive.NoTtl(),
     });
-
-    await removeAllDatabases();
-
+    database = (await encryptionClient.databases.createIfNotExists({ id: randomUUID() })).database;
     const revokedKekMetadata = new EncryptionKeyWrapMetadata(
       testKeyVault,
       "revokedKek",
       "revokedKek-metadata",
       KeyEncryptionKeyAlgorithm.RSA_OAEP,
     );
-    const testdatabase = (await encryptionClient.databases.createIfNotExists({ id: "myDb" }))
-      .database;
-    await testdatabase.createClientEncryptionKey(
-      "keyWithRevokedKek",
-      EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
-      revokedKekMetadata,
+    await testCreateClientEncryptionKey(database, "key1", metadata1);
+    await testCreateClientEncryptionKey(database, "key2", metadata2);
+    await testCreateClientEncryptionKey(database, "keyWithRevokedKek", revokedKekMetadata);
+    const key1Paths = [
+      "/PK",
+      "/sensitive_NestedObjectFormatL1",
+      "/sensitive_IntFormat",
+      "/sensitive_DateFormat",
+      "/sensitive_FloatFormat",
+      "/sensitive_ArrayMultiTypes",
+    ].map(
+      (path) =>
+        new ClientEncryptionIncludedPath(
+          path,
+          "key1",
+          EncryptionType.DETERMINISTIC,
+          EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
+        ),
     );
+    const key2Paths = [
+      "/id",
+      "/sensitive_DecimalFormat",
+      "/sensitive_BoolFormat",
+      "/sensitive_IntMultiDimArray",
+      "/sensitive_ObjectArrayType",
+    ].map(
+      (path) =>
+        new ClientEncryptionIncludedPath(
+          path,
+          "key2",
+          EncryptionType.DETERMINISTIC,
+          EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
+        ),
+    );
+    const paths = key1Paths.concat(key2Paths);
+    paths.push(
+      new ClientEncryptionIncludedPath(
+        "/sensitive_ArrayFormat",
+        "keyWithRevokedKek",
+        EncryptionType.DETERMINISTIC,
+        EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
+      ),
+    );
+    clientEncryptionPolicy = new ClientEncryptionPolicy(paths, 2);
+    containerDefinition = {
+      id: randomUUID(),
+      partitionKey: {
+        paths: ["/PK"],
+      },
+      clientEncryptionPolicy: clientEncryptionPolicy,
+    };
+
+    const containerDefinitionForChangeFeed = {
+      id: randomUUID(),
+      partitionKey: {
+        paths: ["/PK"],
+      },
+      clientEncryptionPolicy: clientEncryptionPolicy,
+    };
+    encryptionContainer = (await database.containers.createIfNotExists(containerDefinition))
+      .container;
+    encryptionContainerForChangeFeed = (
+      await database.containers.createIfNotExists(containerDefinitionForChangeFeed)
+    ).container;
+
+    await encryptionContainer.initializeEncryption();
+    await encryptionContainerForChangeFeed.initializeEncryption();
+  });
+
+  beforeEach(async () => {
+    const iterator = encryptionContainer.items.readAll();
+    const { resources: items } = await iterator.fetchAll();
+    if (items.length > 0) {
+      for (const item of items) {
+        await encryptionContainer.item(item.id, item.PK).delete();
+      }
+    }
+  });
+
+  it("create client encryption included paths and policy", async () => {
+    let path = new ClientEncryptionIncludedPath(
+      "/id",
+      "key1",
+      EncryptionType.DETERMINISTIC,
+      EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
+    );
+    // check policy format version
+    const policy = new ClientEncryptionPolicy([path]);
+    const testcontainer = (
+      await database.containers.createIfNotExists({
+        id: randomUUID(),
+        clientEncryptionPolicy: policy,
+      })
+    ).container;
+    try {
+      await testcontainer.initializeEncryption();
+    } catch (err) {
+      assert.ok(
+        err.message.includes(
+          "Encryption of partition key or id is only supported with policy format version 2",
+        ),
+      );
+    }
+    await testcontainer.delete();
+
+    path = new ClientEncryptionIncludedPath(
+      "id",
+      "key1",
+      EncryptionType.DETERMINISTIC,
+      EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
+    );
+    // check invalid path
+    try {
+      new ClientEncryptionPolicy([path]);
+    } catch (err) {
+      assert.ok(
+        err.message.includes("Path in ClientEncryptionIncludedPath needs to start with '/'"),
+      );
+    }
+    path = new ClientEncryptionIncludedPath(
+      "/id",
+      "",
+      EncryptionType.DETERMINISTIC,
+      EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
+    );
+    // check empty key
+    try {
+      new ClientEncryptionPolicy([path]);
+    } catch (err) {
+      assert.ok(
+        err.message.includes(
+          "ClientEncryptionKeyId needs to be defined in ClientEncryptionIncludedPath.",
+        ),
+      );
+    }
+  });
+
+  it("create and delete database with encryption enabled client without cek", async () => {
+    const testdatabase = (await encryptionClient.databases.create({ id: "NoCEKDatabase" }))
+      .database;
+    const containerProperties = {
+      id: "NoCEPContainer",
+      partitionKey: {
+        paths: ["/PK"],
+      },
+      throughput: 400,
+    };
+    const testcontainer = (await testdatabase.containers.create(containerProperties)).container;
+    await testcontainer.initializeEncryption();
+    const testDoc = TestDoc.create();
+    const createResponse = await testcontainer.items.create(testDoc);
+    assert.equal(StatusCodes.Created, createResponse.statusCode);
+    verifyExpectedDocResponse(testDoc, createResponse.resource);
+    await testdatabase.delete();
+  });
+
+  it("validate encryption time to live", async () => {
+    const testKeyResolver = new MockKeyVaultEncryptionKeyResolver();
+    // client with ttl of 1 min
+    const newClient = new CosmosClient({
+      endpoint: endpoint,
+      key: masterKey,
+      enableEncryption: true,
+      keyEncryptionKeyResolver: testKeyResolver,
+      encryptionKeyResolverName: testKeyVault,
+      encryptionKeyTimeToLive: EncryptionTimeToLive.FromMinutes(1),
+    });
+    const newDatabase = newClient.database(database.id);
+    const newContainer = newDatabase.container(encryptionContainer.id);
+    for (let i = 0; i < 2; i++) {
+      await testCreateItem(newContainer);
+    }
+    // expecting just one unwrap, since the key should be cached for 1 min
+    assert.ok(testKeyResolver.unwrapKeyCallsCount["tempmetadata1"] === 1);
+    // wait for 2 min to ensure that cache is cleared
+    await new Promise((resolve) => setTimeout(resolve, EncryptionTimeToLive.FromMinutes(1)));
+
+    for (let i = 0; i < 2; i++) {
+      await testCreateItem(newContainer);
+    }
+    // again the unwrap key should be called once, since the cache has cleared
+    assert.equal(testKeyResolver.unwrapKeyCallsCount["tempmetadata1"], 2);
+    newClient.dispose();
+  });
+
+  it("encryption bulk operation", async () => {
+    const docToCreate = TestDoc.create();
+
+    const { resource: docToReplace } = await testCreateItem(encryptionContainer);
+    docToReplace.nonsensitive = randomUUID();
+    docToReplace.sensitive_StringFormat = randomUUID();
+
+    const { resource: docToUpsert } = await testCreateItem(encryptionContainer);
+    docToUpsert.nonsensitive = randomUUID();
+    docToUpsert.sensitive_StringFormat = randomUUID();
+
+    // doc not created before
+    const docToUpsert2 = TestDoc.create();
+
+    const { resource: docToDelete } = await testCreateItem(encryptionContainer);
+
+    const clientWithBulk = new CosmosClient({
+      endpoint: endpoint,
+      key: masterKey,
+      enableEncryption: true,
+      keyEncryptionKeyResolver: new MockKeyVaultEncryptionKeyResolver(),
+      encryptionKeyResolverName: testKeyVault,
+    });
+
+    const databaseWithBulk = clientWithBulk.database(database.id);
+    const encryptionContainerWithBulk = databaseWithBulk.container(encryptionContainer.id);
+
+    const operations = [
+      {
+        operationType: BulkOperationType.Create,
+        resourceBody: JSON.parse(JSON.stringify(docToCreate)),
+      },
+      {
+        operationType: BulkOperationType.Upsert,
+        partitionKey: docToUpsert2.PK,
+        resourceBody: JSON.parse(JSON.stringify(docToUpsert2)),
+      },
+      {
+        operationType: BulkOperationType.Replace,
+        partitionKey: docToReplace.PK,
+        id: docToReplace.id,
+        resourceBody: JSON.parse(JSON.stringify(docToReplace)),
+      },
+      {
+        operationType: BulkOperationType.Upsert,
+        partitionKey: docToUpsert.PK,
+        resourceBody: JSON.parse(JSON.stringify(docToUpsert)),
+      },
+      {
+        operationType: BulkOperationType.Delete,
+        id: docToDelete.id,
+        partitionKey: docToDelete.PK,
+      },
+    ];
+
+    const response = await encryptionContainerWithBulk.items.bulk(operations);
+    verifyDiagnostics(response.diagnostics, true, true, undefined, undefined);
+    assert.equal(StatusCodes.Created, response[0].statusCode);
+    verifyExpectedDocResponse(docToCreate, response[0].resourceBody);
+    assert.equal(StatusCodes.Created, response[1].statusCode);
+    verifyExpectedDocResponse(docToUpsert2, response[1].resourceBody);
+    assert.equal(StatusCodes.Ok, response[2].statusCode);
+    verifyExpectedDocResponse(new TestDoc(docToReplace), response[2].resourceBody);
+    assert.equal(StatusCodes.Ok, response[3].statusCode);
+    verifyExpectedDocResponse(new TestDoc(docToUpsert), response[3].resourceBody);
+    assert.equal(StatusCodes.NoContent, response[4].statusCode);
+    assert.isNotObject(response[4].resourceBody);
+    clientWithBulk.dispose();
+  });
+
+  it("encryption create client encryption key", async () => {
+    let cekId = "anotherCek";
+    let testmetadata1 = new EncryptionKeyWrapMetadata(
+      testKeyVault,
+      cekId,
+      "testmetadata1",
+      KeyEncryptionKeyAlgorithm.RSA_OAEP,
+    );
+    const clientEncryptionKeyProperties = (
+      await testCreateClientEncryptionKey(database, cekId, testmetadata1)
+    ).clientEncryptionKeyProperties;
+
+    assert.ok(
+      compareMetadata(testmetadata1, clientEncryptionKeyProperties.encryptionKeyWrapMetadata),
+    );
+
+    // creating another key with same id should fail
+    testmetadata1 = new EncryptionKeyWrapMetadata(
+      testKeyVault,
+      cekId,
+      "testmetadata2",
+      KeyEncryptionKeyAlgorithm.RSA_OAEP,
+    );
+    try {
+      await testCreateClientEncryptionKey(database, cekId, testmetadata1);
+      assert.fail("creating another key with same encryption key id should fail");
+    } catch (error) {
+      assert.equal(StatusCodes.Conflict, error.code);
+    }
+
+    cekId = "testAkvKid";
+    const encryptionCosmosClient = new CosmosClient({
+      endpoint: endpoint,
+      key: masterKey,
+      enableEncryption: true,
+      keyEncryptionKeyResolver: new MockKeyVaultEncryptionKeyResolver(),
+      encryptionKeyResolverName: testKeyVault,
+      encryptionKeyTimeToLive: EncryptionTimeToLive.NoTtl(),
+    });
+    let metadata = new EncryptionKeyWrapMetadata(
+      EncryptionKeyResolverName.AzureKeyVault,
+      "key1",
+      "https://testkeyvault.vault.azure.net/keys/testkey/12345678",
+      KeyEncryptionKeyAlgorithm.RSA_OAEP,
+    );
+    const database1 = (
+      await encryptionCosmosClient.databases.createIfNotExists({ id: randomUUID() })
+    ).database;
+    let clientEncryptionKeyResponse = await database1.createClientEncryptionKey(
+      cekId,
+      EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
+      metadata,
+    );
+    assert.equal(StatusCodes.Created, clientEncryptionKeyResponse.statusCode);
+    metadata = new EncryptionKeyWrapMetadata(
+      EncryptionKeyResolverName.AzureKeyVault,
+      "key1",
+      "https://testkeyvault.vault.azure.net/keys/testkey/9101112",
+      KeyEncryptionKeyAlgorithm.RSA_OAEP,
+    );
+    clientEncryptionKeyResponse = await database1.rewrapClientEncryptionKey(cekId, metadata);
+    assert.equal(StatusCodes.Ok, clientEncryptionKeyResponse.statusCode);
+    encryptionCosmosClient.dispose();
+  });
+
+  it("rewrap client encryption key", async () => {
+    const cekId = "rewrapkeytest";
+    const metadata = new EncryptionKeyWrapMetadata(
+      testKeyVault,
+      cekId,
+      "testmetadata1",
+      KeyEncryptionKeyAlgorithm.RSA_OAEP,
+    );
+    let clientEncryptionKeyProperties = (
+      await testCreateClientEncryptionKey(database, cekId, metadata)
+    ).clientEncryptionKeyProperties;
+    assert.ok(
+      compareMetadata(
+        new EncryptionKeyWrapMetadata(
+          testKeyVault,
+          cekId,
+          metadata.value,
+          KeyEncryptionKeyAlgorithm.RSA_OAEP,
+        ),
+        clientEncryptionKeyProperties.encryptionKeyWrapMetadata,
+      ),
+    );
+    const updatedMetadata = new EncryptionKeyWrapMetadata(
+      testKeyVault,
+      cekId,
+      "metadata" + "updatedmetadata",
+      KeyEncryptionKeyAlgorithm.RSA_OAEP,
+    );
+    clientEncryptionKeyProperties = (
+      await testRewrapClientEncryptionKey(database, cekId, updatedMetadata)
+    ).clientEncryptionKeyProperties;
+    // check if cek is wrapped with updated metadata
+    assert.ok(
+      compareMetadata(
+        new EncryptionKeyWrapMetadata(
+          testKeyVault,
+          cekId,
+          updatedMetadata.value,
+          KeyEncryptionKeyAlgorithm.RSA_OAEP,
+        ),
+        clientEncryptionKeyProperties.encryptionKeyWrapMetadata,
+      ),
+    );
+  });
+
+  it("create encrypted item with null property", async () => {
+    const testkeyEncryptionKeyResolver = new MockKeyVaultEncryptionKeyResolver();
+    const client = new CosmosClient({
+      endpoint: endpoint,
+      key: masterKey,
+      enableEncryption: true,
+      keyEncryptionKeyResolver: testkeyEncryptionKeyResolver,
+      encryptionKeyResolverName: testKeyVault,
+      encryptionKeyTimeToLive: EncryptionTimeToLive.NoTtl(),
+    });
+    const testdatabase = client.database(database.id);
+    const testcontainer = testdatabase.container(encryptionContainer.id);
+    let testDoc = TestDoc.create();
+    testDoc.sensitive_ArrayFormat = null;
+    testDoc.sensitive_StringFormat = null;
+    testDoc.sensitive_NestedObjectFormatL1.sensitive_NestedObjectFormatL2.sensitive_StringFormatL2 =
+      null;
+
+    let createResponse = await testcontainer.items.create(testDoc);
+    assert.equal(StatusCodes.Created, createResponse.statusCode);
+    verifyExpectedDocResponse(testDoc, createResponse.resource);
+
+    // query on document with null property
+
+    const queryBuilder = new EncryptionQueryBuilder(
+      "SELECT * FROM c where c.sensitive_StringFormat = @sensitive_StringFormat AND c.sensitive_ArrayFormat = @sensitive_ArrayFormat" +
+        " AND c.sensitive_IntFormat = @sensitive_IntFormat" +
+        " AND c.sensitive_NestedObjectFormatL1.sensitive_NestedObjectFormatL2.sensitive_StringFormatL2 = @sensitive_StringFormatL2" +
+        " AND c.sensitive_NestedObjectFormatL1.sensitive_NestedObjectFormatL2.sensitive_DecimalFormatL2 = @sensitive_DecimalFormatL2",
+    );
+    queryBuilder.addStringParameter(
+      "@sensitive_StringFormat",
+      testDoc.sensitive_StringFormat,
+      "/sensitive_StringFormat",
+    );
+    queryBuilder.addArrayParameter(
+      "@sensitive_ArrayFormat",
+      testDoc.sensitive_ArrayFormat,
+      "/sensitive_ArrayFormat",
+    );
+    queryBuilder.addIntegerParameter(
+      "@sensitive_IntFormat",
+      testDoc.sensitive_IntFormat,
+      "/sensitive_IntFormat",
+    );
+    queryBuilder.addStringParameter(
+      "@sensitive_StringFormatL2",
+      testDoc.sensitive_NestedObjectFormatL1.sensitive_NestedObjectFormatL2
+        .sensitive_StringFormatL2,
+      "/sensitive_NestedObjectFormatL1",
+    );
+    queryBuilder.addFloatParameter(
+      "@sensitive_DecimalFormatL2",
+      testDoc.sensitive_NestedObjectFormatL1.sensitive_NestedObjectFormatL2
+        .sensitive_DecimalFormatL2,
+      "/sensitive_NestedObjectFormatL1",
+    );
+
+    const expectedDocList: TestDoc[] = [];
+    expectedDocList.push(new TestDoc(testDoc));
+    await validateQueryResults(testcontainer, queryBuilder, expectedDocList);
+    client.dispose();
+  });
+
+  it("encryption create item and query", async () => {
+    const testDoc = (await testCreateItem(encryptionContainer)).resource;
+    await verifyItemByRead(encryptionContainer, testDoc);
+
+    const expectedDoc = new TestDoc(testDoc);
+    const expectedDocList: TestDoc[] = [];
+    expectedDocList.push(expectedDoc);
+    let queryBuilder = new EncryptionQueryBuilder("SELECT * FROM c");
+    await validateQueryResults(encryptionContainer, queryBuilder, expectedDocList);
+    queryBuilder = new EncryptionQueryBuilder(
+      "select * from c where c.id = @theId and c.PK = @thePK",
+    );
+
+    queryBuilder.addStringParameter("@theId", expectedDoc.id, "/id");
+    queryBuilder.addStringParameter("@thePK", expectedDoc.PK, "/PK");
+
+    await validateQueryResults(encryptionContainer, queryBuilder, expectedDocList);
+
+    // query on non encrypted property
+    const querySpec = {
+      query: `SELECT * FROM c WHERE c.nonsensitive = '${expectedDoc.nonsensitive}'`,
+    };
+    await validateQueryResults(encryptionContainer, querySpec, expectedDocList);
+
+    // response should be null without using addIntegerParameter method
+    queryBuilder = new EncryptionQueryBuilder(
+      `SELECT * FROM c WHERE c.sensitive_IntFormat = '${expectedDoc.sensitive_IntFormat}'`,
+    );
+    await validateQueryResults(encryptionContainer, queryBuilder, null);
+  });
+
+  it("query on encrypted properties", async () => {
+    const containerProperties = {
+      id: randomUUID(),
+      partitionKey: {
+        paths: ["/PK"],
+      },
+      clientEncryptionPolicy: clientEncryptionPolicy,
+    };
+    const encryptionQueryContainer = (await database.containers.create(containerProperties))
+      .container;
+    encryptionQueryContainer.initializeEncryption();
+    const testDoc1 = new TestDoc((await testCreateItem(encryptionQueryContainer)).resource);
+    const testDoc2 = new TestDoc((await testCreateItem(encryptionQueryContainer)).resource);
+    const testDoc3 = new TestDoc((await testCreateItem(encryptionQueryContainer)).resource);
+    const arrayOfStringValues = [testDoc1.sensitive_StringFormat, "randomValue", null];
+    let queryBuilder: EncryptionQueryBuilder;
+    // string/object
+    queryBuilder = new EncryptionQueryBuilder(
+      "SELECT * FROM c where array_contains(@sensitive_StringFormat, c.sensitive_StringFormat) AND c.sensitive_NestedObjectFormatL1 = @sensitive_NestedObjectFormatL1",
+    );
+    queryBuilder.addArrayParameter(
+      "@sensitive_StringFormat",
+      arrayOfStringValues,
+      "/sensitive_StringFormat",
+    );
+    queryBuilder.addObjectParameter(
+      "@sensitive_NestedObjectFormatL1",
+      testDoc1.sensitive_NestedObjectFormatL1,
+      "/sensitive_NestedObjectFormatL1",
+    );
+    await validateQueryResults(encryptionQueryContainer, queryBuilder, [testDoc1]);
+    // bool/float
+    queryBuilder = new EncryptionQueryBuilder(
+      "SELECT * FROM c where c.sensitive_BoolFormat = @sensitive_BoolFormat and c.sensitive_FloatFormat = @sensitive_FloatFormat",
+    );
+    queryBuilder.addBooleanParameter(
+      "@sensitive_BoolFormat",
+      testDoc1.sensitive_BoolFormat,
+      "/sensitive_BoolFormat",
+    );
+    queryBuilder.addFloatParameter(
+      "@sensitive_FloatFormat",
+      testDoc1.sensitive_FloatFormat,
+      "/sensitive_FloatFormat",
+    );
+    await validateQueryResults(encryptionQueryContainer, queryBuilder, [
+      testDoc1,
+      testDoc2,
+      testDoc3,
+    ]);
+
+    // with encrypted int and non encrypted properties
+    const testDoc4 = new TestDoc((await testCreateItem(encryptionQueryContainer)).resource);
+    queryBuilder = new EncryptionQueryBuilder(
+      "SELECT * FROM c where c.nonsensitive = @nonsensitive and c.sensitive_IntFormat = @sensitive_IntFormat",
+    );
+    queryBuilder.addStringParameter("@nonsensitive", testDoc4.nonsensitive, "/nonsensitive");
+    queryBuilder.addIntegerParameter(
+      "@sensitive_IntFormat",
+      testDoc4.sensitive_IntFormat,
+      "/sensitive_IntFormat",
+    );
+    await validateQueryResults(encryptionQueryContainer, queryBuilder, [testDoc4]);
+
+    // without adding param
+    queryBuilder = new EncryptionQueryBuilder("SELECT c.sensitive_DateFormat FROM c");
+    await validateQueryResults(encryptionQueryContainer, queryBuilder, null, true, 1);
+  });
+
+  it("encryption batch CRUD", async () => {
+    const partitionKey = "thePK";
+    const doc1ToCreate = TestDoc.create(partitionKey);
+
+    const doc1ToReplace = new TestDoc(
+      (await testCreateItem(encryptionContainer, partitionKey)).resource,
+    );
+    doc1ToReplace.nonsensitive = randomUUID();
+    doc1ToReplace.sensitive_StringFormat = randomUUID();
+
+    const doc2ToReplace = new TestDoc(
+      (await testCreateItem(encryptionContainer, partitionKey)).resource,
+    );
+    doc2ToReplace.nonsensitive = randomUUID();
+    doc2ToReplace.sensitive_StringFormat = randomUUID();
+
+    const doc1ToUpsert = new TestDoc(
+      (await testCreateItem(encryptionContainer, partitionKey)).resource,
+    );
+    doc1ToUpsert.nonsensitive = randomUUID();
+    doc1ToUpsert.sensitive_StringFormat = randomUUID();
+
+    const docToDelete = new TestDoc(
+      (await testCreateItem(encryptionContainer, partitionKey)).resource,
+    );
+
+    const docToPatch = new TestDoc(
+      (await testCreateItem(encryptionContainer, partitionKey)).resource,
+    );
+    docToPatch.sensitive_StringFormat = randomUUID();
+    docToPatch.sensitive_DecimalFormat = 11.11;
+    const patchOperations = {
+      operations: [
+        {
+          op: PatchOperationType.replace,
+          path: "/sensitive_StringFormat",
+          value: docToPatch.sensitive_StringFormat,
+        },
+        {
+          op: PatchOperationType.replace,
+          path: "/sensitive_DecimalFormat",
+          value: docToPatch.sensitive_DecimalFormat,
+        },
+      ],
+    };
+    const operations: OperationInput[] = [
+      {
+        operationType: "Create",
+        resourceBody: JSON.parse(JSON.stringify(doc1ToCreate)),
+      },
+      {
+        operationType: "Upsert",
+        resourceBody: JSON.parse(JSON.stringify(doc1ToUpsert)),
+      },
+      {
+        operationType: "Replace",
+        id: doc1ToReplace.id,
+        resourceBody: JSON.parse(JSON.stringify(doc1ToReplace)),
+      },
+      {
+        operationType: "Patch",
+        id: docToPatch.id,
+        resourceBody: patchOperations,
+      },
+      {
+        operationType: "Delete",
+        id: docToDelete.id,
+      },
+    ];
+
+    const response = await encryptionContainer.items.batch(operations, partitionKey);
+    verifyDiagnostics(response.diagnostics, true, true, undefined, undefined);
+    assert.equal(StatusCodes.Ok, response.code);
+
+    const doc1 = response.result[0];
+
+    verifyExpectedDocResponse(doc1ToCreate, doc1.resourceBody);
+
+    const doc2 = response.result[1];
+    verifyExpectedDocResponse(doc1ToUpsert, doc2.resourceBody);
+
+    const doc3 = response.result[2];
+    verifyExpectedDocResponse(doc1ToReplace, doc3.resourceBody);
+
+    const doc4 = response.result[3];
+    verifyExpectedDocResponse(docToPatch, doc4.resourceBody);
+
+    await verifyItemByRead(encryptionContainer, doc1ToCreate);
+    await verifyItemByRead(encryptionContainer, doc1ToReplace);
+    await verifyItemByRead(encryptionContainer, doc1ToUpsert);
+    await verifyItemByRead(encryptionContainer, docToPatch);
+
+    const readResponse = await encryptionContainer.item(docToDelete.id, docToDelete.PK).read();
+    assert.equal(StatusCodes.NotFound, readResponse.statusCode);
+  });
+
+  it("encryption batch conflict response", async () => {
+    const partitionKey = "thePK";
+    const doc1ToCreateAgain = new TestDoc(
+      (await testCreateItem(encryptionContainer, partitionKey)).resource,
+    );
+    doc1ToCreateAgain.nonsensitive = randomUUID();
+    doc1ToCreateAgain.sensitive_StringFormat = randomUUID();
+
+    const response = await encryptionContainer.items.batch(
+      [
+        {
+          operationType: "Create",
+          resourceBody: JSON.parse(JSON.stringify(doc1ToCreateAgain)),
+        },
+      ],
+      partitionKey,
+    );
+    assert.equal(StatusCodes.Conflict, response.result[0].statusCode);
+    assert.equal(1, response.result.length);
+  });
+
+  it("encryption patch items", async () => {
+    const docPostPatching = new TestDoc((await testCreateItem(encryptionContainer)).resource);
+    docPostPatching.nonsensitive = randomUUID();
+    docPostPatching.nonsensitiveInt++;
+    docPostPatching.sensitive_StringFormat = randomUUID();
+    docPostPatching.sensitive_DateFormat = new Date();
+    docPostPatching.sensitive_DecimalFormat = 11.11;
+    docPostPatching.sensitive_IntArray[1] = 1;
+    docPostPatching.sensitive_IntMultiDimArray[1][0] = 7;
+    docPostPatching.sensitive_IntFormat = 2020;
+    docPostPatching.sensitive_BoolFormat = false;
+    docPostPatching.sensitive_FloatFormat = 2020.2;
+
+    let patchOperations: PatchOperation[] = [
+      {
+        op: PatchOperationType.incr,
+        path: "/nonsensitiveInt",
+        value: 1,
+      },
+      {
+        op: PatchOperationType.replace,
+        path: "/nonsensitive",
+        value: docPostPatching.nonsensitive,
+      },
+      {
+        op: PatchOperationType.replace,
+        path: "/sensitive_StringFormat",
+        value: docPostPatching.sensitive_StringFormat,
+      },
+      {
+        op: PatchOperationType.replace,
+        path: "/sensitive_DateFormat",
+        value: docPostPatching.sensitive_DateFormat,
+      },
+      {
+        op: PatchOperationType.replace,
+        path: "/sensitive_DecimalFormat",
+        value: docPostPatching.sensitive_DecimalFormat,
+      },
+      {
+        op: PatchOperationType.set,
+        path: "/sensitive_IntArray/1",
+        value: docPostPatching.sensitive_IntArray[1],
+      },
+      {
+        op: PatchOperationType.set,
+        path: "/sensitive_IntMultiDimArray/1/0",
+        value: docPostPatching.sensitive_IntMultiDimArray[1][0],
+      },
+      {
+        op: PatchOperationType.replace,
+        path: "/sensitive_IntFormat",
+        value: docPostPatching.sensitive_IntFormat,
+      },
+      {
+        op: PatchOperationType.replace,
+        path: "/sensitive_BoolFormat",
+        value: docPostPatching.sensitive_BoolFormat,
+      },
+      {
+        op: PatchOperationType.replace,
+        path: "/sensitive_FloatFormat",
+        value: docPostPatching.sensitive_FloatFormat,
+      },
+    ];
+    let patchResponse = await testPatchItem(
+      encryptionContainer,
+      patchOperations,
+      docPostPatching,
+      StatusCodes.Ok,
+    );
+    // verifyDiagnostics(patchResponse.diagnostics, true, true, 6);
+    docPostPatching.sensitive_ArrayFormat = [
+      {
+        sensitive_ArrayIntFormat: 100,
+        sensitive_ArrayDecimalFormat: 100.2,
+      },
+      {
+        sensitive_ArrayIntFormat: 2222,
+        sensitive_ArrayDecimalFormat: 2222.22,
+      },
+    ];
+    docPostPatching.sensitive_ArrayMultiTypes[0][1] = {
+      sensitive_NestedObjectFormatL0: {
+        sensitive_IntFormatL0: 123,
+        sensitive_DecimalFormatL0: 123.1,
+      },
+      sensitive_StringArrayMultiType: ["sensitivedata"],
+      sensitive_ArrayMultiTypeDecimalFormat: 123.2,
+      sensitive_IntArrayMultiType: [1, 2],
+    };
+    docPostPatching.sensitive_NestedObjectFormatL1 = {
+      sensitive_IntArrayL1: [999, 100],
+      sensitive_IntFormatL1: 1999,
+      sensitive_DecimalFormatL1: 1.1,
+      sensitive_ArrayFormatL1: [
+        {
+          sensitive_ArrayIntFormat: 0,
+          sensitive_ArrayDecimalFormat: 0.1,
+        },
+        {
+          sensitive_ArrayIntFormat: 0,
+          sensitive_ArrayDecimalFormat: 0.5,
+        },
+        {
+          sensitive_ArrayIntFormat: 1,
+          sensitive_ArrayDecimalFormat: 2.1,
+        },
+        {
+          sensitive_ArrayIntFormat: 2,
+          sensitive_ArrayDecimalFormat: 3.1,
+        },
+      ],
+    };
+    // patch operation on nested path properties
+    patchOperations = [
+      {
+        op: PatchOperationType.replace,
+        path: "/sensitive_ArrayFormat/0",
+        value: docPostPatching.sensitive_ArrayFormat[0],
+      },
+      {
+        op: PatchOperationType.add,
+        path: "/sensitive_ArrayFormat/1",
+        value: docPostPatching.sensitive_ArrayFormat[1],
+      },
+      {
+        op: PatchOperationType.replace,
+        path: "/sensitive_ArrayMultiTypes/0/1",
+        value: docPostPatching.sensitive_ArrayMultiTypes[0][1],
+      },
+      {
+        op: PatchOperationType.remove,
+        path: "/sensitive_NestedObjectFormatL1/sensitive_NestedObjectFormatL2",
+      },
+      {
+        op: PatchOperationType.set,
+        path: "/sensitive_NestedObjectFormatL1/sensitive_ArrayFormatL1/0",
+        value: docPostPatching.sensitive_NestedObjectFormatL1.sensitive_ArrayFormatL1[0],
+      },
+      {
+        op: PatchOperationType.set,
+        path: "/sensitive_NestedObjectFormatL1/sensitive_ArrayFormatL1/1",
+        value: docPostPatching.sensitive_NestedObjectFormatL1.sensitive_ArrayFormatL1[1],
+      },
+      {
+        op: PatchOperationType.replace,
+        path: "/sensitive_NestedObjectFormatL1/sensitive_DecimalFormatL1",
+        value: docPostPatching.sensitive_NestedObjectFormatL1.sensitive_DecimalFormatL1,
+      },
+    ];
+    patchResponse = await testPatchItem(
+      encryptionContainer,
+      patchOperations,
+      docPostPatching,
+      StatusCodes.Ok,
+    );
+    verifyDiagnostics(patchResponse.diagnostics, true, true, 6, 12);
+    patchOperations = [
+      {
+        op: PatchOperationType.incr,
+        path: "/sensitive_IntFormat",
+        value: 1,
+      },
+    ];
+    try {
+      await encryptionContainer.item(docPostPatching.id, docPostPatching.PK).patch(patchOperations);
+    } catch (err) {
+      assert.ok(
+        err.message.includes(
+          "Increment patch operation is not allowed for encrypted path '/sensitive_IntFormat'",
+        ),
+      );
+    }
+  });
+
+  it("Encryption RUD item", async () => {
+    // upsert should create a new item
+    const testDoc = new TestDoc(
+      (await testUpsertItem(encryptionContainer, TestDoc.create(), StatusCodes.Created)).resource,
+    );
+    await verifyItemByRead(encryptionContainer, testDoc);
+    testDoc.nonsensitive = randomUUID();
+    testDoc.sensitive_StringFormat = randomUUID();
+
+    // upsert should update the item
+    const updatedDoc = new TestDoc(
+      (await testUpsertItem(encryptionContainer, testDoc, StatusCodes.Ok)).resource,
+    );
+    await verifyItemByRead(encryptionContainer, updatedDoc);
+
+    updatedDoc.nonsensitive = randomUUID();
+    updatedDoc.sensitive_StringFormat = randomUUID();
+
+    const replacedDoc = new TestDoc(
+      (await testReplaceItem(encryptionContainer, updatedDoc)).resource,
+    );
+    await verifyItemByRead(encryptionContainer, replacedDoc);
+    await testDeleteItem(encryptionContainer, replacedDoc);
+  });
+
+  it("validate Partition Key and ID encryption support", async () => {
+    // encrypt string type PK
+    let paths = ["/PK", "/id"].map(
+      (path) =>
+        new ClientEncryptionIncludedPath(
+          path,
+          "key1",
+          EncryptionType.DETERMINISTIC,
+          EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
+        ),
+    );
+    let clientEncryptionPolicy = new ClientEncryptionPolicy(paths, 2);
+    let containerProperties = {
+      id: "StringPKEncContainer",
+      partitionKey: {
+        paths: ["/id"],
+      },
+      clientEncryptionPolicy: clientEncryptionPolicy,
+    };
+    let testcontainer = (await database.containers.create(containerProperties)).container;
+    await testcontainer.initializeEncryption();
+    const testDoc = TestDoc.create();
+    let createResponse = await testcontainer.items.create(testDoc);
+    assert.equal(StatusCodes.Created, createResponse.statusCode);
+    verifyExpectedDocResponse(testDoc, createResponse.resource);
+    // readback
+    let readResponse = await testcontainer.item(testDoc.id, testDoc.id).read();
+    assert.equal(StatusCodes.Ok, readResponse.statusCode);
+    verifyExpectedDocResponse(testDoc, readResponse.resource);
+
+    // encrypt float type PK
+    paths = ["/sensitive_FloatFormat", "/id"].map(
+      (path) =>
+        new ClientEncryptionIncludedPath(
+          path,
+          "key1",
+          EncryptionType.DETERMINISTIC,
+          EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
+        ),
+    );
+    clientEncryptionPolicy = new ClientEncryptionPolicy(paths, 2);
+    containerProperties = {
+      id: "FloatPKEncContainer",
+      partitionKey: {
+        paths: ["/sensitive_FloatFormat"],
+      },
+      clientEncryptionPolicy: clientEncryptionPolicy,
+    };
+    testcontainer = (await database.containers.create(containerProperties)).container;
+    await testcontainer.initializeEncryption();
+
+    createResponse = await testcontainer.items.create(testDoc);
+    assert.equal(StatusCodes.Created, createResponse.statusCode);
+    verifyExpectedDocResponse(testDoc, createResponse.resource);
+
+    // readback
+    readResponse = await testcontainer.item(testDoc.id, testDoc.sensitive_FloatFormat).read();
+    assert.equal(StatusCodes.Ok, readResponse.statusCode);
+    verifyExpectedDocResponse(testDoc, readResponse.resource);
+
+    // encrypt boolean type PK
+    paths = ["/sensitive_BoolFormat", "/id"].map(
+      (path) =>
+        new ClientEncryptionIncludedPath(
+          path,
+          "key1",
+          EncryptionType.DETERMINISTIC,
+          EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
+        ),
+    );
+    clientEncryptionPolicy = new ClientEncryptionPolicy(paths, 2);
+    containerProperties = {
+      id: "BoolPKEncContainer",
+      partitionKey: {
+        paths: ["/sensitive_BoolFormat"],
+      },
+      clientEncryptionPolicy: clientEncryptionPolicy,
+    };
+    testcontainer = (await database.containers.create(containerProperties)).container;
+    await testcontainer.initializeEncryption();
+
+    createResponse = await testcontainer.items.create(testDoc);
+    assert.equal(StatusCodes.Created, createResponse.statusCode);
+    verifyExpectedDocResponse(testDoc, createResponse.resource);
+    // readback
+    readResponse = await testcontainer.item(testDoc.id, testDoc.sensitive_BoolFormat).read();
+    assert.equal(StatusCodes.Ok, readResponse.statusCode);
+    verifyExpectedDocResponse(testDoc, readResponse.resource);
+  });
+
+  it("create and delete database with encryption enabled client without cek", async () => {
+    const testdatabase = (await encryptionClient.databases.create({ id: "NoCEKDatabase" }))
+      .database;
+    const containerProperties = {
+      id: "NoCEPContainer",
+      partitionKey: {
+        paths: ["/PK"],
+      },
+    };
+    const testcontainer = (await testdatabase.containers.create(containerProperties)).container;
+    await testcontainer.initializeEncryption();
+    const testDoc = TestDoc.create();
+    const createResponse = await testcontainer.items.create(testDoc);
+    assert.equal(StatusCodes.Created, createResponse.statusCode);
+    verifyExpectedDocResponse(testDoc, createResponse.resource);
+    await testdatabase.delete();
+  });
+
+  // tests non-encrypted create operation with encrypted client
+  it("encryption create item with no client encryption policy", async () => {
+    await testCreateItem(encryptionContainer);
+    const containerProperties = {
+      id: randomUUID(),
+      partitionKey: {
+        paths: ["/PK"],
+      },
+    };
+    const encryptionContainerWithNoPolicy = (await database.containers.create(containerProperties))
+      .container;
+    await encryptionContainerWithNoPolicy.initializeEncryption();
+
+    const testDoc = TestDoc.create();
+    const createResponse = await encryptionContainerWithNoPolicy.items.create(testDoc);
+    assert.equal(StatusCodes.Created, createResponse.statusCode);
+    verifyExpectedDocResponse(testDoc, createResponse.resource);
+
+    const queryBuilder = new EncryptionQueryBuilder(
+      `SELECT * FROM c where c.sensitive_StringFormat = @sensitive_StringFormat AND c.sensitive_IntFormat = @sensitive_IntFormat`,
+    );
+    queryBuilder.addStringParameter(
+      "@sensitive_StringFormat",
+      testDoc.sensitive_StringFormat,
+      "/sensitive_StringFormat",
+    );
+    queryBuilder.addIntegerParameter(
+      "@sensitive_IntFormat",
+      testDoc.sensitive_IntFormat,
+      "/sensitive_IntFormat",
+    );
+    const expectedDocList = [testDoc];
+    await validateQueryResults(
+      encryptionContainerWithNoPolicy,
+      queryBuilder,
+      expectedDocList,
+      false,
+    );
+    await encryptionContainerWithNoPolicy.delete();
+  });
+
+  it("encryption change feed", async () => {
+    const partitionKey = randomUUID();
+    const changeFeedOptionsForEntireContainer = {
+      changeFeedStartFrom: ChangeFeedStartFrom.Beginning(),
+      maxItemCount: 1,
+    };
+    const changeFeedOptionsForPartitionKey = {
+      changeFeedStartFrom: ChangeFeedStartFrom.Beginning(partitionKey),
+    };
+
+    const iteratorForPartitionKey = encryptionContainer.items.getChangeFeedIterator(
+      changeFeedOptionsForPartitionKey,
+    );
+    const iteratorForEntireContainer = encryptionContainer.items.getChangeFeedIterator(
+      changeFeedOptionsForEntireContainer,
+    );
+    const testDoc1 = new TestDoc(
+      (await testCreateItem(encryptionContainer, partitionKey)).resource,
+    );
+    const testDoc2 = new TestDoc((await testCreateItem(encryptionContainer)).resource);
+
+    let response = await iteratorForEntireContainer.readNext();
+
+    verifyExpectedDocResponse(testDoc1, response.result[0]);
+    assert.equal(StatusCodes.Ok, response.statusCode);
+    response = await iteratorForEntireContainer.readNext();
+    verifyExpectedDocResponse(testDoc2, response.result[0]);
+
+    response = await iteratorForPartitionKey.readNext();
+    assert.equal(StatusCodes.Ok, response.statusCode);
+    assert.ok(response.result.length === 1);
+    verifyExpectedDocResponse(testDoc1, response.result[0]);
+  });
+
+  it("encryption change feed with allVersionsAndDeletes", async () => {
+    const newClient = new CosmosClient({
+      endpoint: endpoint,
+      key: masterKey,
+      enableEncryption: true,
+      keyEncryptionKeyResolver: new MockKeyVaultEncryptionKeyResolver(),
+      encryptionKeyResolverName: testKeyVault,
+    });
+    const newDatabase = newClient.database(database.id);
+    const newTimeStamp = ChangeFeedRetentionTimeSpan.fromMinutes(5);
+    const changeFeedPolicy = new ChangeFeedPolicy(newTimeStamp);
+    const containerDefinitionChangeFeed = {
+      id: randomUUID(),
+      partitionKey: {
+        paths: ["/PK"],
+        version: PartitionKeyDefinitionVersion.V1,
+      },
+      clientEncryptionPolicy: clientEncryptionPolicy,
+      changeFeedPolicy: changeFeedPolicy,
+    };
+
+    const encryptionContainerChangeFeed = (
+      await newDatabase.containers.createIfNotExists(containerDefinitionChangeFeed)
+    ).container;
+
+    const partitionKey = randomUUID();
+    // test encryption change feed with allVersionsAndDeletes for entire container
+    const changeFeedOptionsForEntireContainer = {
+      changeFeedStartFrom: ChangeFeedStartFrom.Now(),
+      maxItemCount: 1,
+      changeFeedMode: ChangeFeedMode.AllVersionsAndDeletes,
+    };
+
+    const iteratorForEntireContainer = encryptionContainerChangeFeed.items.getChangeFeedIterator(
+      changeFeedOptionsForEntireContainer,
+    );
+    await iteratorForEntireContainer.readNext();
+    const testDoc1 = new TestDoc(
+      (await testCreateItem(encryptionContainerChangeFeed, partitionKey)).resource,
+    );
+    const response = await iteratorForEntireContainer.readNext();
+    assert.equal(StatusCodes.Ok, response.statusCode);
+    if ("current" in response.result[0]) {
+      verifyExpectedDocResponse(testDoc1, response.result[0].current);
+    } else {
+      assert.fail("current not found in response");
+    }
+    verifyDiagnostics(response.diagnostics, false, true, 12, 12);
+    // test encryption change feed with allVersionsAndDeletes for partition key
+    const changeFeedOptionsForPartitionKey = {
+      changeFeedStartFrom: ChangeFeedStartFrom.Now(partitionKey),
+      changeFeedMode: ChangeFeedMode.AllVersionsAndDeletes,
+    };
+
+    const iteratorForPartitionKey = encryptionContainerChangeFeed.items.getChangeFeedIterator(
+      changeFeedOptionsForPartitionKey,
+    );
+    await iteratorForPartitionKey.readNext();
+    const testDoc2 = new TestDoc(
+      (await testCreateItem(encryptionContainerChangeFeed, partitionKey)).resource,
+    );
+    await testCreateItem(encryptionContainerChangeFeed);
+
+    await testDeleteItem(encryptionContainerChangeFeed, testDoc1);
+    const responseForPartitionKey = await iteratorForPartitionKey.readNext();
+    assert.equal(StatusCodes.Ok, responseForPartitionKey.statusCode);
+    assert.ok(responseForPartitionKey.count === 2);
+
+    if ("current" in responseForPartitionKey.result[0]) {
+      verifyExpectedDocResponse(testDoc2, responseForPartitionKey.result[0].current);
+    } else {
+      assert.fail("current not found in response");
+    }
+    if ("previous" in responseForPartitionKey.result[1]) {
+      verifyExpectedDocResponse(testDoc1, responseForPartitionKey.result[1].previous);
+    } else {
+      assert.fail("previous not found in response");
+    }
+  });
+
+  it("encryption validate policy refresh post container delete with patch", async () => {
+    // create a container with 1st client
+    let paths = [
+      "/sensitive_IntArray",
+      "/sensitive_NestedObjectFormatL1",
+      "/sensitive_DoubleFormat",
+    ].map(
+      (path) =>
+        new ClientEncryptionIncludedPath(
+          path,
+          "key1",
+          EncryptionType.DETERMINISTIC,
+          EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
+        ),
+    );
+    let encryptionPolicy = new ClientEncryptionPolicy(paths, 2);
+    let containerProperties: ContainerDefinition = {
+      id: randomUUID(),
+      partitionKey: { paths: ["/sensitive_DoubleFormat"] },
+      clientEncryptionPolicy: encryptionPolicy,
+    };
+    const encryptionContainerToDelete = (await database.containers.create(containerProperties))
+      .container;
+    await encryptionContainerToDelete.initializeEncryption();
+    // create a document with 2nd client on same database and container
+    const otherClient = new CosmosClient({
+      endpoint: endpoint,
+      key: masterKey,
+      enableEncryption: true,
+      keyEncryptionKeyResolver: new MockKeyVaultEncryptionKeyResolver(),
+      encryptionKeyResolverName: testKeyVault,
+    });
+    const otherDatabase = otherClient.database(database.id);
+    const otherEncryptionContainer = otherDatabase.container(encryptionContainerToDelete.id);
+    const testDoc = TestDoc.create();
+    const createResponse = await otherEncryptionContainer.items.create(testDoc);
+    assert.equal(StatusCodes.Created, createResponse.statusCode);
+    verifyExpectedDocResponse(testDoc, createResponse.resource);
+    // Client 1 Deletes the Container referenced in Client 2 and Recreate with different policy
+    await database.container(encryptionContainerToDelete.id).delete();
+    paths = [
+      new ClientEncryptionIncludedPath(
+        "/sensitive_StringFormat",
+        "key1",
+        EncryptionType.DETERMINISTIC,
+        EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
+      ),
+      new ClientEncryptionIncludedPath(
+        "/sensitive_BoolFormat",
+        "key2",
+        EncryptionType.DETERMINISTIC,
+        EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
+      ),
+      new ClientEncryptionIncludedPath(
+        "/PK",
+        "key2",
+        EncryptionType.DETERMINISTIC,
+        EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
+      ),
+    ];
+    encryptionPolicy = new ClientEncryptionPolicy(paths, 2);
+    containerProperties = {
+      id: encryptionContainerToDelete.id,
+      partitionKey: { paths: ["/PK"] },
+      clientEncryptionPolicy: encryptionPolicy,
+    };
+    await database.containers.create(containerProperties);
+    try {
+      await testCreateItem(encryptionContainerToDelete);
+      assert.fail("create operation should fail");
+    } catch (err) {
+      verifyDiagnostics(err.diagnostics, true, false, 3, 0);
+      assert.ok(
+        err.message.includes(
+          "Operation has failed due to a possible mismatch in Client Encryption Policy configured on the container.",
+        ),
+      );
+    }
+
+    const docPostPatching = (await testCreateItem(encryptionContainerToDelete)).resource;
+    docPostPatching.nonsensitive = randomUUID();
+    docPostPatching.nonSensitiveInt++;
+    docPostPatching.sensitive_StringFormat = randomUUID();
+    docPostPatching.sensitive_DateFormat = new Date();
+    docPostPatching.sensitive_IntArray[1] = 19877;
+    docPostPatching.sensitive_IntMultiDimArray[1][0] = 2020;
+    docPostPatching.sensitive_IntFormat = 2020;
+    docPostPatching.sensitive_NestedObjectFormatL1 = {
+      sensitive_IntArrayL1: [999, 100],
+      sensitive_IntFormatL1: 1999,
+      sensitive_DecimalFormatL1: 1991.1,
+      sensitive_ArrayFormatL1: [
+        {
+          sensitive_ArrayIntFormat: 0,
+          sensitive_ArrayDecimalFormat: 0.1,
+        },
+        {
+          sensitive_ArrayIntFormat: 1,
+          sensitive_ArrayDecimalFormat: 2.1,
+        },
+        {
+          sensitive_ArrayIntFormat: 2,
+          sensitive_ArrayDecimalFormat: 3.1,
+        },
+      ],
+    };
+    const patchOperations = [
+      {
+        op: PatchOperationType.incr,
+        path: "/nonSensitiveInt",
+        value: 1,
+      },
+      {
+        op: PatchOperationType.replace,
+        path: "/nonsensitive",
+        value: docPostPatching.nonsensitive,
+      },
+      {
+        op: PatchOperationType.replace,
+        path: "/sensitive_StringFormat",
+        value: docPostPatching.sensitive_StringFormat,
+      },
+      {
+        op: PatchOperationType.replace,
+        path: "/sensitive_DateFormat",
+        value: docPostPatching.sensitive_DateFormat,
+      },
+      {
+        op: PatchOperationType.set,
+        path: "/sensitive_IntArray/1",
+        value: docPostPatching.sensitive_IntArray[1],
+      },
+      {
+        op: PatchOperationType.set,
+        path: "/sensitive_IntMultiDimArray/1/0",
+        value: docPostPatching.sensitive_IntMultiDimArray[1][0],
+      },
+      {
+        op: PatchOperationType.replace,
+        path: "/sensitive_IntFormat",
+        value: docPostPatching.sensitive_IntFormat,
+      },
+      {
+        op: PatchOperationType.remove,
+        path: "/sensitive_NestedObjectFormatL1/sensitive_NestedObjectFormatL2",
+        value: docPostPatching.sensitive_NestedObjectFormatL1.sensitive_NestedObjectFormatL2,
+      },
+      {
+        op: PatchOperationType.set,
+        path: "/sensitive_NestedObjectFormatL1/sensitive_ArrayFormatL1/0",
+        value: docPostPatching.sensitive_NestedObjectFormatL1.sensitive_ArrayFormatL1[0],
+      },
+    ];
+    try {
+      await testPatchItem(
+        otherEncryptionContainer,
+        patchOperations,
+        docPostPatching,
+        StatusCodes.Ok,
+      );
+      assert.fail("patch operation should fail");
+    } catch (err) {
+      verifyDiagnostics(err.diagnostics, true, false, 2, 0);
+      assert.ok(
+        err.message.includes(
+          "Operation has failed due to a possible mismatch in Client Encryption Policy configured on the container.",
+        ),
+      );
+    }
+    await testPatchItem(otherEncryptionContainer, patchOperations, docPostPatching, StatusCodes.Ok);
+    otherClient.dispose();
+  });
+
+  it("encryption validate policy refresh post container delete with query", async () => {
+    // create a container with 1st client
+    let paths = [
+      "/sensitive_IntArray",
+      "/sensitive_NestedObjectFormatL1",
+      "/sensitive_DoubleFormat",
+    ].map(
+      (path) =>
+        new ClientEncryptionIncludedPath(
+          path,
+          "key1",
+          EncryptionType.DETERMINISTIC,
+          EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
+        ),
+    );
+    let encryptionPolicy = new ClientEncryptionPolicy(paths, 2);
+    let containerProperties: ContainerDefinition = {
+      id: randomUUID(),
+      partitionKey: { paths: ["/sensitive_DoubleFormat"] },
+      clientEncryptionPolicy: encryptionPolicy,
+    };
+    const encryptionContainerToDelete = (await database.containers.create(containerProperties))
+      .container;
+    await encryptionContainerToDelete.initializeEncryption();
+    // create a document with 2nd client on same database and container
+    const otherClient = new CosmosClient({
+      endpoint: endpoint,
+      key: masterKey,
+      enableEncryption: true,
+      keyEncryptionKeyResolver: new MockKeyVaultEncryptionKeyResolver(),
+      encryptionKeyResolverName: testKeyVault,
+    });
+    const otherDatabase = otherClient.database(database.id);
+    const otherEncryptionContainer = otherDatabase.container(encryptionContainerToDelete.id);
+    const testDoc = TestDoc.create();
+    let createResponse = await otherEncryptionContainer.items.create(testDoc);
+    assert.equal(StatusCodes.Created, createResponse.statusCode);
+    verifyExpectedDocResponse(testDoc, createResponse.resource);
+    // Client 1 Deletes the Container referenced in Client 2 and Recreate with different policy
+    await database.container(encryptionContainerToDelete.id).delete();
+    paths = [
+      new ClientEncryptionIncludedPath(
+        "/sensitive_StringFormat",
+        "key1",
+        EncryptionType.DETERMINISTIC,
+        EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
+      ),
+      new ClientEncryptionIncludedPath(
+        "/sensitive_BoolFormat",
+        "key2",
+        EncryptionType.DETERMINISTIC,
+        EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
+      ),
+      new ClientEncryptionIncludedPath(
+        "/PK",
+        "key2",
+        EncryptionType.DETERMINISTIC,
+        EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
+      ),
+    ];
+    encryptionPolicy = new ClientEncryptionPolicy(paths, 2);
+    containerProperties = {
+      id: encryptionContainerToDelete.id,
+      partitionKey: { paths: ["/PK"] },
+      clientEncryptionPolicy: encryptionPolicy,
+    };
+    await database.containers.create(containerProperties);
+    try {
+      await testCreateItem(encryptionContainerToDelete);
+      assert.fail("create operation should fail");
+    } catch (err) {
+      verifyDiagnostics(err.diagnostics, true, false, 3, 0);
+      assert.ok(
+        err.message.includes(
+          "Operation has failed due to a possible mismatch in Client Encryption Policy configured on the container.",
+        ),
+      );
+    }
+    createResponse = await encryptionContainerToDelete.items.create(testDoc);
+    assert.equal(StatusCodes.Created, createResponse.statusCode);
+    verifyExpectedDocResponse(testDoc, createResponse.resource);
+
+    // check w.r.t to query if we are able to fail and update the policy
+    try {
+      await validateQueryResults(
+        otherEncryptionContainer,
+        new EncryptionQueryBuilder("SELECT * FROM c"),
+        [testDoc],
+        true,
+        3,
+      );
+      assert.fail("query operation should fail");
+    } catch (err) {
+      assert.ok(
+        err.message.includes(
+          "Operation has failed due to a possible mismatch in Client Encryption Policy configured on the container.",
+        ),
+      );
+    }
+    // previous failure would have updated the policy in the cache.
+    await validateQueryResults(
+      otherEncryptionContainer,
+      new EncryptionQueryBuilder("SELECT * FROM c"),
+      [testDoc],
+      true,
+      3,
+    );
+    otherClient.dispose();
+  });
+
+  it("encryption validate policy refresh post container delete with batch", async () => {
+    // create a container with 1st client
+    let paths = [
+      "/sensitive_IntArray",
+      "/sensitive_NestedObjectFormatL1",
+      "/sensitive_DoubleFormat",
+    ].map(
+      (path) =>
+        new ClientEncryptionIncludedPath(
+          path,
+          "key1",
+          EncryptionType.DETERMINISTIC,
+          EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
+        ),
+    );
+    let encryptionPolicy = new ClientEncryptionPolicy(paths, 2);
+    let containerProperties: ContainerDefinition = {
+      id: randomUUID(),
+      partitionKey: { paths: ["/sensitive_DoubleFormat"] },
+      clientEncryptionPolicy: encryptionPolicy,
+    };
+    const encryptionContainerToDelete = (await database.containers.create(containerProperties))
+      .container;
+    await encryptionContainerToDelete.initializeEncryption();
+    // create a document with 2nd client on same database and container
+    const otherClient = new CosmosClient({
+      endpoint: endpoint,
+      key: masterKey,
+      enableEncryption: true,
+      keyEncryptionKeyResolver: new MockKeyVaultEncryptionKeyResolver(),
+      encryptionKeyResolverName: testKeyVault,
+    });
+    const otherDatabase = otherClient.database(database.id);
+    const otherEncryptionContainer = otherDatabase.container(encryptionContainerToDelete.id);
+    let testDoc = TestDoc.create();
+    const createResponse = await otherEncryptionContainer.items.create(testDoc);
+    assert.equal(StatusCodes.Created, createResponse.statusCode);
+    verifyExpectedDocResponse(testDoc, createResponse.resource);
+    // Client 1 Deletes the Container referenced in Client 2 and Recreate with different policy
+    await database.container(encryptionContainerToDelete.id).delete();
+    paths = [
+      new ClientEncryptionIncludedPath(
+        "/sensitive_StringFormat",
+        "key1",
+        EncryptionType.DETERMINISTIC,
+        EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
+      ),
+      new ClientEncryptionIncludedPath(
+        "/sensitive_BoolFormat",
+        "key2",
+        EncryptionType.DETERMINISTIC,
+        EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
+      ),
+      new ClientEncryptionIncludedPath(
+        "/PK",
+        "key2",
+        EncryptionType.DETERMINISTIC,
+        EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
+      ),
+    ];
+    encryptionPolicy = new ClientEncryptionPolicy(paths, 2);
+    containerProperties = {
+      id: encryptionContainerToDelete.id,
+      partitionKey: { paths: ["/PK"] },
+      clientEncryptionPolicy: encryptionPolicy,
+    };
+    await database.containers.create(containerProperties);
+    try {
+      await testCreateItem(encryptionContainerToDelete);
+      assert.fail("create operation should fail");
+    } catch (err) {
+      verifyDiagnostics(err.diagnostics, true, false, 3, 0);
+      assert.ok(
+        err.message.includes(
+          "Operation has failed due to a possible mismatch in Client Encryption Policy configured on the container.",
+        ),
+      );
+    }
+    testDoc = new TestDoc((await testCreateItem(encryptionContainerToDelete)).resource);
+    const partitionKey = "thePK";
+
+    const doc1ToCreate = TestDoc.create(partitionKey);
+    const doc2ToCreate = TestDoc.create(partitionKey);
+    // check w.r.t to Batch if we are able to fail and update the policy.
+    const operations: OperationInput[] = [
+      {
+        operationType: "Create",
+        resourceBody: JSON.parse(JSON.stringify(doc1ToCreate)),
+      },
+      {
+        operationType: "Create",
+        resourceBody: JSON.parse(JSON.stringify(doc2ToCreate)),
+      },
+      {
+        operationType: "Read",
+        id: doc1ToCreate.id,
+      },
+      {
+        operationType: "Delete",
+        id: doc2ToCreate.id,
+      },
+    ];
+    let batchResponse = null;
+    try {
+      batchResponse = await otherEncryptionContainer.items.batch(operations, partitionKey);
+      assert.fail("batch operation should fail");
+    } catch (err) {
+      assert.ok(
+        err.message.includes(
+          "Operation has failed due to a possible mismatch in Client Encryption Policy configured on the container.",
+        ),
+      );
+    }
+    batchResponse = await otherEncryptionContainer.items.batch(operations, partitionKey);
+    assert.equal(StatusCodes.Ok, batchResponse.code);
+    const doc1 = batchResponse.result[0].resourceBody;
+    verifyExpectedDocResponse(doc1ToCreate, doc1);
+    const doc2 = batchResponse.result[1].resourceBody;
+    verifyExpectedDocResponse(doc2ToCreate, doc2);
+    otherClient.dispose();
+  });
+
+  it("encryption validate policy refresh post container delete with bulk", async () => {
+    // create a container with 1st client
+    let paths = [
+      "/sensitive_IntArray",
+      "/sensitive_NestedObjectFormatL1",
+      "/sensitive_DoubleFormat",
+    ].map(
+      (path) =>
+        new ClientEncryptionIncludedPath(
+          path,
+          "key1",
+          EncryptionType.DETERMINISTIC,
+          EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
+        ),
+    );
+    let encryptionPolicy = new ClientEncryptionPolicy(paths, 2);
+    let containerProperties: ContainerDefinition = {
+      id: randomUUID(),
+      partitionKey: { paths: ["/sensitive_DoubleFormat"] },
+      clientEncryptionPolicy: encryptionPolicy,
+    };
+    const encryptionContainerToDelete = (await database.containers.create(containerProperties))
+      .container;
+    await encryptionContainerToDelete.initializeEncryption();
+    // create a document with 2nd client on same database and container
+    const otherClient = new CosmosClient({
+      endpoint: endpoint,
+      key: masterKey,
+      enableEncryption: true,
+      keyEncryptionKeyResolver: new MockKeyVaultEncryptionKeyResolver(),
+      encryptionKeyResolverName: testKeyVault,
+    });
+    const otherDatabase = otherClient.database(database.id);
+    const otherEncryptionContainer = otherDatabase.container(encryptionContainerToDelete.id);
+    const testDoc = TestDoc.create();
+    const createResponse = await otherEncryptionContainer.items.create(testDoc);
+    assert.equal(StatusCodes.Created, createResponse.statusCode);
+    verifyExpectedDocResponse(testDoc, createResponse.resource);
+    // Client 1 Deletes the Container referenced in Client 2 and Recreate with different policy
+    await database.container(encryptionContainerToDelete.id).delete();
+    paths = [
+      new ClientEncryptionIncludedPath(
+        "/sensitive_StringFormat",
+        "key1",
+        EncryptionType.DETERMINISTIC,
+        EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
+      ),
+      new ClientEncryptionIncludedPath(
+        "/sensitive_BoolFormat",
+        "key2",
+        EncryptionType.DETERMINISTIC,
+        EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
+      ),
+      new ClientEncryptionIncludedPath(
+        "/PK",
+        "key2",
+        EncryptionType.DETERMINISTIC,
+        EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
+      ),
+    ];
+    encryptionPolicy = new ClientEncryptionPolicy(paths, 2);
+    containerProperties = {
+      id: encryptionContainerToDelete.id,
+      partitionKey: { paths: ["/PK"] },
+      clientEncryptionPolicy: encryptionPolicy,
+    };
+    await database.containers.create(containerProperties);
+    try {
+      await testCreateItem(encryptionContainerToDelete);
+      assert.fail("create operation should fail");
+    } catch (err) {
+      verifyDiagnostics(err.diagnostics, true, false, 3, 3);
+      assert.ok(
+        err.message.includes(
+          "Operation has failed due to a possible mismatch in Client Encryption Policy configured on the container.",
+        ),
+      );
+    }
+    const docToReplace = (await testCreateItem(encryptionContainerToDelete)).resource;
+    docToReplace.sensitive_StringFormat = "docToBeReplaced";
+    const docToUpsert = (await testCreateItem(encryptionContainerToDelete)).resource;
+    docToUpsert.sensitive_StringFormat = "docToBeUpserted";
+    const operations = [
+      {
+        operationType: BulkOperationType.Upsert,
+        partitionKey: docToUpsert.PK,
+        resourceBody: JSON.parse(JSON.stringify(docToUpsert)),
+      },
+      {
+        operationType: BulkOperationType.Replace,
+        id: docToReplace.id,
+        partitionKey: docToReplace.PK,
+        resourceBody: JSON.parse(JSON.stringify(docToReplace)),
+      },
+      {
+        operationType: BulkOperationType.Create,
+        resourceBody: JSON.parse(JSON.stringify(TestDoc.create())),
+      },
+    ];
+    try {
+      await otherEncryptionContainer.items.bulk(operations);
+      assert.fail("bulk operation should fail");
+    } catch (error) {
+      verifyDiagnostics(error.diagnostics, true, false, undefined, undefined);
+      assert.ok(
+        error.message.includes(
+          "Operation has failed due to a possible mismatch in Client Encryption Policy configured on the container.",
+        ),
+      );
+    }
+    // retry bulk operation with 2nd client
+    const res = await otherEncryptionContainer.items.bulk(operations);
+    assert.equal(StatusCodes.Ok, res[0].statusCode);
+    assert.equal(StatusCodes.Ok, res[1].statusCode);
+    assert.equal(StatusCodes.Created, res[2].statusCode);
+    await verifyItemByRead(encryptionContainerToDelete, docToReplace);
+    await testCreateItem(encryptionContainerToDelete);
+    await verifyItemByRead(encryptionContainerToDelete, docToUpsert);
+    // validate if the right policy was used, by reading them all back
+    await otherEncryptionContainer.items.readAll().fetchAll();
+    otherClient.dispose();
+  });
+
+  it("encryption validate policy refresh post database delete", async () => {
+    const mainCLient = new CosmosClient({
+      endpoint: endpoint,
+      key: masterKey,
+      enableEncryption: true,
+      keyEncryptionKeyResolver: new MockKeyVaultEncryptionKeyResolver(),
+      encryptionKeyResolverName: testKeyVault,
+      encryptionKeyTimeToLive: EncryptionTimeToLive.FromMinutes(30),
+    });
+    let keyWrapMetadata = new EncryptionKeyWrapMetadata(
+      testKeyVault,
+      "myCek",
+      "mymetadata1",
+      KeyEncryptionKeyAlgorithm.RSA_OAEP,
+    );
+    let mainDatabase = (await mainCLient.databases.createIfNotExists({ id: "databaseToBeDeleted" }))
+      .database;
+    await mainDatabase.createClientEncryptionKey(
+      keyWrapMetadata.name,
+      EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
+      keyWrapMetadata,
+    );
+
+    const originalPaths = [
+      "/sensitive_StringFormat",
+      "/sensitive_ArrayFormat",
+      "/sensitive_NestedObjectFormatL1",
+    ].map(
+      (path) =>
+        new ClientEncryptionIncludedPath(
+          path,
+          "myCek",
+          EncryptionType.DETERMINISTIC,
+          EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
+        ),
+    );
+    const encryptionPolicy = new ClientEncryptionPolicy(originalPaths);
+    let containerDef: ContainerDefinition = {
+      id: "containerToBeDeleted",
+      partitionKey: {
+        paths: ["/PK"],
+      },
+      clientEncryptionPolicy: encryptionPolicy,
+    };
+    const encryptionContainerToDelete = (
+      await mainDatabase.containers.createIfNotExists(containerDef)
+    ).container;
+    await encryptionContainerToDelete.initializeEncryption();
+    await testCreateItem(encryptionContainerToDelete);
+
+    const otherClient = new CosmosClient({
+      endpoint: endpoint,
+      key: masterKey,
+      enableEncryption: true,
+      keyEncryptionKeyResolver: new MockKeyVaultEncryptionKeyResolver(),
+      encryptionKeyResolverName: testKeyVault,
+      encryptionKeyTimeToLive: EncryptionTimeToLive.NoTtl(),
+    });
+    const otherDatabase = otherClient.database(mainDatabase.id);
+    const otherEncryptionContainer = otherDatabase.container(encryptionContainerToDelete.id);
+    await testCreateItem(otherEncryptionContainer);
+
+    // Client 1 Deletes the Database and  Container referenced in Client 2 and Recreate with different policy
+    // delete database and recreate with same key name
+    await mainDatabase.delete();
+    mainDatabase = (await mainCLient.databases.createIfNotExists({ id: "databaseToBeDeleted" }))
+      .database;
+    keyWrapMetadata = new EncryptionKeyWrapMetadata(
+      testKeyVault,
+      "myCek",
+      "mymetadata2",
+      KeyEncryptionKeyAlgorithm.RSA_OAEP,
+    );
+    await mainDatabase.createClientEncryptionKey(
+      keyWrapMetadata.name,
+      EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
+      keyWrapMetadata,
+    );
+    const newModifiedPaths = [
+      "/sensitive_IntArray",
+      "/sensitive_DateFormat",
+      "/sensitive_BoolFormat",
+    ].map(
+      (path) =>
+        new ClientEncryptionIncludedPath(
+          path,
+          "myCek",
+          EncryptionType.DETERMINISTIC,
+          EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
+        ),
+    );
+    clientEncryptionPolicy = new ClientEncryptionPolicy(newModifiedPaths);
+    containerDef = {
+      id: encryptionContainerToDelete.id,
+      partitionKey: {
+        paths: ["/PK"],
+      },
+      clientEncryptionPolicy: clientEncryptionPolicy,
+    };
+    await mainDatabase.containers.createIfNotExists(containerDef);
+
+    try {
+      await testCreateItem(encryptionContainerToDelete);
+      assert.fail("create operation should fail");
+    } catch (error) {
+      error.message.includes(
+        "Operation has failed due to a possible mismatch in Client Encryption Policy configured on the container.",
+      );
+    }
+    // retrying should succeed
+    let testDoc = new TestDoc((await testCreateItem(encryptionContainerToDelete)).resource);
+
+    // try from other container. should fail due to policy mismatch
+    try {
+      await verifyItemByRead(otherEncryptionContainer, testDoc);
+      assert.fail("read operation should fail");
+    } catch (error) {
+      error.message.includes(
+        "Operation has failed due to a possible mismatch in Client Encryption Policy configured on the container.",
+      );
+    }
+    // retrying should succeed
+    await verifyItemByRead(otherEncryptionContainer, testDoc);
+
+    // create new container in other client.
+    // The test basically validates if the new key created is referenced, Since the other client would have had the old key cached.
+    // and here we would not hit the incorrect container rid issue.
+    const newModifiedPath2 = new ClientEncryptionIncludedPath(
+      "/PK",
+      "myCek",
+      EncryptionType.DETERMINISTIC,
+      EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
+    );
+    const otherEncryptionContainer2 = (
+      await otherDatabase.containers.createIfNotExists({
+        id: "otherContainer2",
+        partitionKey: "/PK",
+        clientEncryptionPolicy: new ClientEncryptionPolicy([newModifiedPath2], 2),
+      })
+    ).container;
+
+    // create an item
+    const newDoc = new TestDoc((await testCreateItem(otherEncryptionContainer2)).resource);
+    const otherClient2 = new CosmosClient({
+      endpoint: endpoint,
+      key: masterKey,
+      enableEncryption: true,
+      keyEncryptionKeyResolver: new MockKeyVaultEncryptionKeyResolver(),
+      encryptionKeyResolverName: testKeyVault,
+      encryptionKeyTimeToLive: EncryptionTimeToLive.FromHours(1),
+    });
+    const otherDatabase2 = otherClient2.database(mainDatabase.id);
+    const otherEncryptionContainer3 = otherDatabase2.container(otherEncryptionContainer2.id);
+    await verifyItemByRead(otherEncryptionContainer3, newDoc);
+
+    // validate from other client that we indeed are using the key with metadata2
+    const otherEncryptionContainerFromClient2 = otherDatabase2.container(
+      encryptionContainerToDelete.id,
+    );
+    await verifyItemByRead(otherEncryptionContainerFromClient2, testDoc);
+
+    // previous referenced container
+    await testCreateItem(otherEncryptionContainer);
+
+    await testCreateItem(otherEncryptionContainer);
+    await verifyItemByRead(otherEncryptionContainer, testDoc);
+
+    testDoc = new TestDoc((await testCreateItem(otherEncryptionContainer)).resource);
+
+    // to be sure if it was indeed encrypted with the new key.
+    await verifyItemByRead(encryptionContainerToDelete, testDoc);
+
+    // validate if the right policy was used, by reading them all back
+    await encryptionContainerToDelete.items.readAll().fetchAll();
+    await otherEncryptionContainer.items.readAll().fetchAll();
+
+    await mainCLient.database(mainDatabase.id).delete();
+
+    mainCLient.dispose();
+    otherClient.dispose();
+    otherClient2.dispose();
+  });
+
+  it("encryption decrypt group by query result test", async () => {
+    const partitionKey = randomUUID();
+    const testDoc1 = new TestDoc((await testCreateItem(encryptionContainer)).resource);
+
+    const query = new EncryptionQueryBuilder(
+      `SELECT COUNT(c.id), c.PK FROM c WHERE c.PK = @PK GROUP BY c.PK`,
+    );
+
+    query.addStringParameter("@PK", partitionKey, "/PK");
+    let iterator = await encryptionContainer.items.getEncryptionQueryIterator(query);
+    while (iterator.hasMoreResults()) {
+      const response = await iterator.fetchNext();
+      assert.isNotNull(response.diagnostics);
+    }
+
+    const withEncryptedParameter = new EncryptionQueryBuilder(
+      "SELECT COUNT(c.id), c.sensitive_IntFormat FROM c WHERE c.sensitive_IntFormat = @Sensitive_IntFormat GROUP BY c.sensitive_IntFormat",
+    );
+
+    withEncryptedParameter.addIntegerParameter(
+      "@Sensitive_IntFormat",
+      testDoc1.sensitive_IntFormat,
+      "/sensitive_IntFormat",
+    );
+
+    iterator = await encryptionContainer.items.getEncryptionQueryIterator(withEncryptedParameter);
+    while (iterator.hasMoreResults()) {
+      const response = await iterator.fetchNext();
+      assert.isNotNull(response.diagnostics);
+    }
+  });
+
+  it("should fail creating cep with duplicate path", async () => {
+    // duplicate paths in policy
+    const pathdup1 = new ClientEncryptionIncludedPath(
+      "/sensitive_StringFormat",
+      "key2",
+      EncryptionType.DETERMINISTIC,
+      EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
+    );
+    const pathdup2 = new ClientEncryptionIncludedPath(
+      "/sensitive_StringFormat",
+      "key2",
+      EncryptionType.DETERMINISTIC,
+      EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
+    );
+    const pathsWithDups = [pathdup1, pathdup2];
+    try {
+      new ClientEncryptionPolicy(pathsWithDups);
+      assert.fail("ClientEncryptionPolicy creation should have failed");
+    } catch (err) {
+      assert.ok(
+        err.message.includes(
+          err.message,
+          "Duplicate path found: /sensitive_StringFormat in client encryption policy.",
+        ),
+      );
+    }
+  });
+
+  it("encryption decrypt query result multiple docs", async () => {
+    const testDoc1 = new TestDoc((await testCreateItem(encryptionContainer)).resource);
+    const testDoc2 = new TestDoc((await testCreateItem(encryptionContainer)).resource);
+
+    let query = `SELECT * FROM c WHERE c.nonsensitive IN ('${testDoc1.nonsensitive}', '${testDoc2.nonsensitive}')`;
+
+    let iterator = encryptionContainer.items.query(query);
+    let response = await iterator.fetchAll();
+    assert.ok(response.resources.length === 2);
+    for (const doc of response.resources) {
+      assert.ok(doc.id === testDoc1.id || doc.id === testDoc2.id);
+    }
+    query += " ORDER BY c._ts";
+    iterator = encryptionContainer.items.query(query);
+    response = await iterator.fetchAll();
+    assert.ok(response.resources.length === 2);
+  });
+
+  it("query with COUNT on encrypted item", async () => {
+    await testCreateItem(encryptionContainer);
+    const query = { query: "SELECT VALUE COUNT(1) FROM c" };
+    const iterator = encryptionContainer.items.query(query);
+    const response = await iterator.fetchAll();
+    assert.equal(response.resources[0], 1);
+  });
+
+  it("encryption hierarchical partition key test", async () => {
+    const key1Paths = ["/sensitive_LongFormat", "/sensitive_NestedObjectFormatL1"].map(
+      (path) =>
+        new ClientEncryptionIncludedPath(
+          path,
+          "key1",
+          EncryptionType.DETERMINISTIC,
+          EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
+        ),
+    );
+    const containerDef = {
+      id: "hierarchical_partition_container",
+      partitionKey: {
+        paths: [
+          "/sensitive_StringFormat",
+          "/sensitive_NestedObjectFormatL1/sensitive_NestedObjectFormatL2/sensitive_StringFormatL2",
+        ],
+        version: 2,
+        kind: PartitionKeyKind.MultiHash,
+      },
+      clientEncryptionPolicy: new ClientEncryptionPolicy(key1Paths, 2),
+      throughput: 400,
+    };
+    const container = (await database.containers.create(containerDef)).container;
+    await container.initializeEncryption();
+    const testDoc = TestDoc.create();
+    const createResponse = await container.items.create(testDoc);
+    assert.equal(StatusCodes.Created, createResponse.statusCode);
+    const testDocPartitionKey = [
+      testDoc.sensitive_StringFormat,
+      testDoc.sensitive_NestedObjectFormatL1.sensitive_NestedObjectFormatL2
+        .sensitive_StringFormatL2,
+    ];
+    verifyExpectedDocResponse(testDoc, createResponse.resource);
+    const readResponse = await container.item(testDoc.id, testDocPartitionKey).read();
+    assert.equal(StatusCodes.Ok, readResponse.statusCode);
+    verifyExpectedDocResponse(testDoc, readResponse.resource);
+    let query = new EncryptionQueryBuilder("SELECT * FROM c");
+    let iterator = await container.items.getEncryptionQueryIterator(query);
+    let queryResponse = await iterator.fetchAll();
+    assert.equal(queryResponse.resources.length, 1);
+    verifyExpectedDocResponse(testDoc, queryResponse.resources[0]);
+    query = new EncryptionQueryBuilder(
+      "SELECT * FROM c WHERE c.sensitive_StringFormat= @sensitive_StringFormat",
+    );
+    query.addStringParameter(
+      "@sensitive_StringFormat",
+      testDoc.sensitive_StringFormat,
+      "/sensitive_StringFormat",
+    );
+    iterator = await container.items.getEncryptionQueryIterator(query);
+    queryResponse = await iterator.fetchAll();
+    assert.equal(queryResponse.resources.length, 1);
+    verifyExpectedDocResponse(testDoc, queryResponse.resources[0]);
+
+    query = new EncryptionQueryBuilder(
+      "SELECT * FROM c WHERE c.sensitive_LongFormat= @sensitive_LongFormat",
+    );
+    query.addIntegerParameter(
+      "@sensitive_LongFormat",
+      testDoc.sensitive_LongFormat,
+      "/sensitive_LongFormat",
+    );
+    iterator = await container.items.getEncryptionQueryIterator(query);
+    queryResponse = await iterator.fetchAll();
+    assert.equal(queryResponse.resources.length, 1);
+    verifyExpectedDocResponse(testDoc, queryResponse.resources[0]);
+
+    query = new EncryptionQueryBuilder(
+      "SELECT * FROM c WHERE c.sensitive_NestedObjectFormatL1= @sensitive_NestedObjectFormatL1",
+    );
+    query.addObjectParameter(
+      "@sensitive_NestedObjectFormatL1",
+      testDoc.sensitive_NestedObjectFormatL1,
+      "/sensitive_NestedObjectFormatL1",
+    );
+    iterator = await container.items.getEncryptionQueryIterator(query);
+    queryResponse = await iterator.fetchAll();
+    assert.equal(queryResponse.resources.length, 1);
+    verifyExpectedDocResponse(testDoc, queryResponse.resources[0]);
+  });
+
+  it("encryption resource token auth restricted", async () => {
+    const { resource: restrictedUserDef } = await database.users.create({ id: randomUUID() });
+    const restrictedUser = database.user(restrictedUserDef.id);
+    const { resource: restrictedUserPermission } = await restrictedUser.permissions.create({
+      id: randomUUID(),
+      permissionMode: PermissionMode.All,
+      resource: encryptionContainer.url,
+    });
+    assert((restrictedUserPermission as any)._token !== undefined, "permission token is invalid");
+    const resourceTokens: any = {};
+    resourceTokens[database.id] = (restrictedUserPermission as any)._token;
+    const restrictedClient = new CosmosClient({
+      endpoint: endpoint,
+      resourceTokens: resourceTokens,
+      enableEncryption: true,
+      keyEncryptionKeyResolver: new MockKeyVaultEncryptionKeyResolver(),
+      encryptionKeyResolverName: testKeyVault,
+    });
+
+    const datbaseForRestrictedUser = restrictedClient.database(database.id);
+
+    try {
+      const cekId = "testingCekId";
+      const metadata = new EncryptionKeyWrapMetadata(
+        testKeyVault,
+        cekId,
+        "testmetadata1",
+        KeyEncryptionKeyAlgorithm.RSA_OAEP,
+      );
+      await datbaseForRestrictedUser.createClientEncryptionKey(
+        cekId,
+        EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
+        metadata,
+      );
+      assert.fail("CreateClientEncryptionKey should have failed due to restrictions");
+    } catch (err) {
+      assert.equal(StatusCodes.Forbidden, err.code);
+    }
+
+    try {
+      const cekId = "testingCekId";
+      const metadata = new EncryptionKeyWrapMetadata(
+        testKeyVault,
+        cekId,
+        "testmetadata1" + "updated",
+        KeyEncryptionKeyAlgorithm.RSA_OAEP,
+      );
+      await datbaseForRestrictedUser.rewrapClientEncryptionKey(cekId, metadata);
+      assert.fail("RewrapClientEncryptionKey should have failed due to restrictions");
+    } catch (err) {
+      assert.equal(StatusCodes.Forbidden, err.code);
+    }
+  });
+
+  it("key encryption key revoke test", async () => {
+    const keyEncryptionKeyResolver = new MockKeyVaultEncryptionKeyResolver();
+    const encryptionTestClient = new CosmosClient({
+      endpoint: endpoint,
+      key: masterKey,
+      enableEncryption: true,
+      keyEncryptionKeyResolver: keyEncryptionKeyResolver,
+      encryptionKeyTimeToLive: EncryptionTimeToLive.NoTtl(),
+      encryptionKeyResolverName: testKeyVault,
+    });
+    const testdatabase = encryptionTestClient.database(database.id);
 
     // Once a Dek gets cached and the Kek is revoked, calls to unwrap/wrap keys would fail since KEK is revoked.
     // The Dek should be rewrapped if the KEK is revoked.
@@ -735,18 +2167,11 @@ describe("Client Side Encryption", () => {
       clientEncryptionPolicy: clientEncryptionPolicyWithRevokedKek,
     };
     const testcontainer = (await testdatabase.containers.create(containerProperties)).container;
+    const testDoc1 = new TestDoc((await testCreateItem(testcontainer)).resource);
     keyEncryptionKeyResolver.revokeAccessSet = true;
     // Try creating it and it should fail as it has been revoked.
     try {
-      await testcontainer.items.create({
-        PK: "1",
-        id: "1",
-        sensitive_NestedObjectFormatL1: {
-          sensitive_NestedObjectFormatL2: {
-            sensitive_StringFormat: "test",
-          },
-        },
-      });
+      await testCreateItem(testcontainer);
       assert.fail("Create Item should have failed.");
     } catch (err) {
       assert.ok(
@@ -755,57 +2180,109 @@ describe("Client Side Encryption", () => {
         ),
       );
     }
+    // testing query read fail due to revoked access.
+    const query = new EncryptionQueryBuilder("SELECT * FROM c");
+    try {
+      await validateQueryResults(testcontainer, query, [testDoc1]);
+      assert.fail("Query should have failed.");
+    } catch (error) {
+      assert.ok(
+        error.message.includes(
+          "needs to be rewrapped with a valid Key Encryption Key using rewrapClientEncryptionKey",
+        ),
+      );
+    }
+    // Revoke access is set to false, so the next call should succeed after rewrap.
+    keyEncryptionKeyResolver.revokeAccessSet = false;
+    await testdatabase.rewrapClientEncryptionKey("keyWithRevokedKek", metadata2);
+    await testCreateItem(testcontainer);
+    testKeyEncryptionKeyResolver.revokeAccessSet = false;
+    encryptionTestClient.dispose();
   });
-  it.skip("test validate caching of protected dek", async () => {
-    const newTestKeyResolver = new MockKeyVaultEncryptionKeyResolver();
-    const newClient = new CosmosClient({
+
+  it("encryption diagnostics test", async () => {
+    const createResponse = await testCreateItem(encryptionContainer);
+    verifyDiagnostics(createResponse.diagnostics, true, true, 12, 12);
+
+    const testDoc = new TestDoc(createResponse.resource);
+
+    const readResponse = await encryptionContainer.item(testDoc.id, testDoc.PK).read();
+    verifyDiagnostics(readResponse.diagnostics, false, true, undefined, 12);
+
+    const testDoc1 = TestDoc.create();
+    testDoc1.nonsensitive = randomUUID();
+    testDoc1.sensitive_StringFormat = randomUUID();
+    const upsertResponse = await testUpsertItem(encryptionContainer, testDoc1, StatusCodes.Created);
+    const upsertedDoc = new TestDoc(upsertResponse.resource);
+    verifyDiagnostics(upsertResponse.diagnostics, true, true, 12, 12);
+
+    upsertedDoc.nonsensitive = randomUUID();
+    upsertedDoc.sensitive_StringFormat = randomUUID();
+
+    const replaceResponse = await testReplaceItem(encryptionContainer, upsertedDoc);
+    verifyDiagnostics(replaceResponse.diagnostics, true, true, 12, 12);
+  });
+
+  it("encryption delete all items in a partition key", async () => {
+    const testDoc1 = new TestDoc((await testCreateItem(encryptionContainer, "pk1")).resource);
+    const testDoc2 = new TestDoc((await testCreateItem(encryptionContainer, "pk2")).resource);
+    const testDoc3 = new TestDoc((await testCreateItem(encryptionContainer, "pk1")).resource);
+    const testDoc4 = new TestDoc((await testCreateItem(encryptionContainer, "pk2")).resource);
+
+    await encryptionContainer.deleteAllItemsForPartitionKey("pk1");
+
+    const res1 = await encryptionContainer.item(testDoc1.id, "pk1").read();
+    assert(res1.statusCode === StatusCodes.NotFound);
+
+    const res2 = await encryptionContainer.item(testDoc2.id, "pk2").read();
+    assert(res2.statusCode === StatusCodes.Ok);
+
+    const res3 = await encryptionContainer.item(testDoc3.id, "pk1").read();
+    assert(res3.statusCode === StatusCodes.NotFound);
+
+    const res4 = await encryptionContainer.item(testDoc4.id, "pk2").read();
+    assert(res4.statusCode === StatusCodes.Ok);
+  });
+
+  it("validate caching of protected dek", async () => {
+    // no caching
+    const testKeyResolver1 = new MockKeyVaultEncryptionKeyResolver();
+    let newClient = new CosmosClient({
       endpoint: endpoint,
       key: masterKey,
       enableEncryption: true,
-      keyEncryptionKeyResolver: newTestKeyResolver,
-      encryptionKeyTimeToLiveInHours: 0,
+      keyEncryptionKeyResolver: testKeyResolver1,
+      encryptionKeyResolverName: testKeyVault,
+      encryptionKeyTimeToLive: EncryptionTimeToLive.NoTtl(),
+    });
+    let newDatabase = newClient.database(database.id);
+    let newContainer = newDatabase.container(encryptionContainer.id);
+    for (let i = 0; i < 2; i++) {
+      await testCreateItem(newContainer);
+    }
+    let unwrapCount = testKeyResolver1.unwrapKeyCallsCount["tempmetadata1"];
+    assert.ok(unwrapCount > 1);
+    const testKeyResolver2 = new MockKeyVaultEncryptionKeyResolver();
+    // default cache ttl is 2 hrs
+    newClient = new CosmosClient({
+      endpoint: endpoint,
+      key: masterKey,
+      enableEncryption: true,
+      keyEncryptionKeyResolver: testKeyResolver2,
       encryptionKeyResolverName: testKeyVault,
     });
-    const newDatabase = newClient.database(database.id);
-    const path = new ClientEncryptionIncludedPath(
-      "/sensitive_NestedObjectFormatL1",
-      "dek1",
-      EncryptionType.DETERMINISTIC,
-      EncryptionAlgorithm.AEAD_AES_256_CBC_HMAC_SHA256,
-    );
-    const clientEncryptionPolicyWithRevokedKek = new ClientEncryptionPolicy([path]);
-    const containerProperties = {
-      id: randomUUID(),
-      partitionKey: {
-        paths: ["/PK"],
-      },
-      clientEncryptionPolicy: clientEncryptionPolicyWithRevokedKek,
-    };
-    const newContainer = (await newDatabase.containers.createIfNotExists(containerProperties))
-      .container;
-    console.log("START 2");
-    await newContainer.items.create({
-      PK: "1",
-      id: "1",
-      sensitive_NestedObjectFormatL1: {
-        sensitive_NestedObjectFormatL2: {
-          sensitive_StringFormat: "test",
-        },
-      },
-    });
-
-    await newContainer.items.create({
-      PK: "2",
-      id: "2",
-      sensitive_NestedObjectFormatL1: {
-        sensitive_NestedObjectFormatL2: {
-          sensitive_StringFormat: "test2",
-        },
-      },
-    });
-    const unwrapcount = newTestKeyResolver.unwrapKeyCallsCount["tempmetadata1"];
-    console.log(JSON.stringify(newTestKeyResolver));
-
-    assert.equal(1, unwrapcount);
+    newDatabase = newClient.database(database.id);
+    newContainer = newDatabase.container(encryptionContainer.id);
+    for (let i = 0; i < 2; i++) {
+      await testCreateItem(newContainer);
+    }
+    unwrapCount = testKeyResolver2.unwrapKeyCallsCount["tempmetadata1"];
+    // expecting just one unwrap
+    assert.ok(unwrapCount === 1);
+    newClient.dispose();
+  });
+  after(async () => {
+    await removeAllDatabases();
+    encryptionClient.dispose();
   });
 });
